@@ -1,13 +1,13 @@
 {
   "name": "sigma-file-manager",
   "productName": "Sigma file manager",
-  "version": "1.1.0",
+  "version": "1.4.0",
   "private": true,
   "license": "GPL-3.0-or-later",
   "author": "Aleksey Hoffman",
   "email": "aleksey.hoffman.public@gmail.com",
   "homepage": "https://github.com/aleksey-hoffman/sigma-file-manager",
-  "description": "\"Sigma file manager\" is a free, open-source, quickly evolving, modern file manager (explorer / finder) app for Windows, MacOS, and Linux.",
+  "description": "\"Sigma File Manager\" is a free, open-source, quickly evolving, modern file manager (explorer / finder) app for Windows, MacOS, and Linux.",
   "keywords": [
     "open-source",
     "app",
@@ -24,10 +24,9 @@
   ],
   "main": "background.js",
   "scripts": {
-    "serve": "vue-cli-service serve",
     "lint": "vue-cli-service lint",
-    "electron:build": "node ./src/prepareFilesForBuild.js && vue-cli-service electron:build",
-    "electron:serve": "node ./src/prepareFilesForDev.js && vue-cli-service electron:serve",
+    "electron:build": "vue-cli-service electron:build",
+    "electron:serve": "vue-cli-service electron:serve",
     "postinstall": "electron-builder install-app-deps",
     "postuninstall": "electron-builder install-app-deps",
     "nibble": "eslint-nibble *"
@@ -38,39 +37,36 @@
     "@mdi/font": "^5.9.55",
     "7zip-bin": "^5.1.1",
     "atomically": "^1.7.0",
-    "byline": "^5.0.0",
-    "chokidar": "^3.5.1",
-    "core-js": "^3.12.1",
-    "dayjs": "^1.10.4",
-    "diskusage": "^1.1.3",
-    "eloquent-ffmpeg": "^0.14.0",
+    "chokidar": "^3.5.2",
+    "core-js": "^3.14.0",
+    "dayjs": "^1.10.5",
+    "diskusage": "github:aleksey-hoffman/node-diskusage",
     "express": "^4.17.1",
     "express-fileupload": "^1.2.1",
     "fast-sort": "^3.0.2",
     "fluent-ffmpeg": "^2.1.2",
     "fs-extra": "^10.0.0",
+    "fswin": "^3.21.905",
     "get-port": "^5.1.1",
-    "hash-wasm": "^4.6.0",
+    "hash-wasm": "^4.7.0",
     "idle-js": "^1.2.0",
-    "katex": "^0.13.10",
+    "katex": "^0.13.11",
     "micromatch": "^4.0.4",
-    "mime-types": "^2.1.30",
+    "mime-types": "^2.1.31",
     "mousetrap": "^1.6.5",
     "node-7z": "^2.1.2",
     "overlayscrollbars": "^1.13.1",
     "overlayscrollbars-vue": "^0.2.2",
     "qrcode": "^1.4.4",
     "readdirp": "^3.6.0",
-    "readline": "^1.3.0",
-    "recursive-copy": "^2.0.11",
     "serve-index": "^1.9.1",
-    "systeminformation": "^5.6.20",
+    "systeminformation": "^5.7.4",
     "through2": "^4.0.2",
     "trammel": "^5.0.0",
     "vue": "^2.6.12",
     "vue-router": "^3.5.1",
     "vue-smooth-dnd": "^0.8.1",
-    "vuetify": "^2.5.0",
+    "vuetify": "^2.5.3",
     "vuex": "^3.6.2",
     "vuex-map-fields": "^1.4.1",
     "xxhash-addon": "^1.4.0"
@@ -83,7 +79,8 @@
     "@vue/cli-service": "~4.5.12",
     "@vue/eslint-config-standard": "^6.0.0",
     "babel-eslint": "^10.1.0",
-    "electron": "^12.0.7",
+    "copy-webpack-plugin": "^5.1.2",
+    "electron": "^12.1.0",
     "eslint": "^7.22.0",
     "eslint-plugin-import": "^2.21.2",
     "eslint-plugin-node": "^11.1.0",
