@@ -4,8 +4,12 @@ Copyright © 2021 - present Aleksey Hoffman. All rights reserved.
 -->
 
 <template>
-  <v-app :data-theme-type="themeType">
-    <!-- global-components -->
+  <v-app
+    :data-theme-type="themeType" 
+    :route-name="$route.name" 
+    :dir-item-background="dirItemBackground"
+    :is-window-maximized="windowsMainStateIsMaximized"
+  >
     <window-toolbar/>
     <action-toolbar/>
     <navigation-panel/>
@@ -13,48 +17,41 @@ Copyright © 2021 - present Aleksey Hoffman. All rights reserved.
     <overlays/>
     <clipboard-toolbar/>
     <dialogs v-if="appIsLoaded"/>
+    <window-effects/>
+    <fs-local-server-manager/>
+    <context-menus/>
 
     <!-- app-content-area -->
     <v-main class="app-content">
-      <transition
-        name="route-transition"
-        mode="out-in"
+      <keep-alive 
+        :include="['home', 'settings']"
       >
-        <keep-alive :include="['home']">
-          <router-view/>
-        </keep-alive>
-      </transition>
+        <router-view/>
+      </keep-alive>
     </v-main>
   </v-app>
 </template>
 
 <script>
 console.time('time::App.vue::Imports')
-import { mapGetters } from 'vuex'
-import { mapFields } from 'vuex-map-fields'
-import { Walk } from './utils/driveWalker.js'
+import {mapGetters} from 'vuex'
+import {mapFields} from 'vuex-map-fields'
+import {DriveWalker} from './utils/driveWalker.js'
 import GlobalSearchWorker from 'worker-loader!./workers/globalSearchWorker.js'
 import DirWatcherWorker from 'worker-loader!./workers/dirWatcherWorker.js'
 import TimeUtils from './utils/timeUtils.js'
-import getStorageDevices from './utils/storageInfo.js'
+import {getStorageDevices} from './utils/storageInfo.js'
 import idleJs from 'idle-js'
+import * as notifications from './utils/notifications.js'
+
 const electron = require('electron')
 const PATH = require('path')
 const fs = require('fs')
+const fsExtra = require('fs-extra')
+const fsInfo = require('./utils/fsInfo.js')
 const mousetrap = require('mousetrap')
-const diskusage = require('diskusage')
-const sysInfo = require('systeminformation')
-const getDirSize = require('trammel')
-const fileUpload = require('express-fileupload')
-const serveIndex = require('serve-index')
-const getPort = require('get-port')
-const net = require('net')
-const qrCode = require('qrcode')
 const childProcess = require('child_process')
 const zlib = require('zlib')
-const node7z = require('node-7z')
-let express
-let expressServer
 console.timeEnd('time::App.vue::Imports')
 
 export default {
@@ -63,8 +60,14 @@ export default {
     $route (to, from) {
       this.contextMenus.dirItem.value = false
       this.$store.dispatch('TERMINATE_ALL_FETCH_DIR_SIZE')
+      // Unload items to improve UI responsiveness
+      this.$store.state.navigatorView.dirItems = []
       if (to.name === 'home') {
-        this.animateHomeBanner({ delay: true })
+        this.preventHomeViewLayoutTransition()
+        this.animateHomeBannerIn()
+      }
+      if (from.name === 'home') {
+        this.animateHomeBannerOut()
       }
     },
     drives (value) {
@@ -81,10 +84,32 @@ export default {
       }
     },
     'inputState.shift' (value) {
-      if (!value) {
+      if (value) {
+        if (this.inputState.pointer.hover.itemType = 'dirItem') {
+          this.$store.dispatch('HANDLE_HIGHLIGHT_DIR_ITEM_RANGE', {
+            hoveredItem: this.inputState.pointer.hover.item
+          })
+        }  
+      }
+      else {
         this.$store.dispatch('DEHIGHLIGHT_ALL_DIR_ITEMS')
       }
-    }
+    },
+    themeType () {
+      this.setCSSAttributes('visual-filters')
+    },
+    visualFiltersApplyFiltersToMediaElements () {
+      this.setCSSAttributes('visual-filters')
+    },
+    visualFiltersContrastValue () {
+      this.setCSSAttributes('visual-filters')
+    },
+    visualFiltersBrightnessValue () {
+      this.setCSSAttributes('visual-filters')
+    },
+    visualFiltersSaturationValue () {
+      this.setCSSAttributes('visual-filters')
+    },
   },
   created () {
     this.$store.dispatch('CLONE_STATE')
@@ -96,7 +121,6 @@ export default {
   async mounted () {
     this.$store.dispatch('ADD_ACTION_TO_HISTORY', { action: 'App.vue::mounted()' })
     try {
-      await this.initMediaDirectories()
       await this.initAllStorageFiles()
       await this.fetchStorageDevices()
       this.handleFirstAppLaunch()
@@ -115,6 +139,7 @@ export default {
       this.checkForAppUpdateInstalled()
       this.initDirWatcherWorker()
       this.initEventHubListeners()
+      electron.ipcRenderer.invoke('main-window-loaded')
     }
     catch (error) {
       electron.ipcRenderer.send('show:errorWindow', {
@@ -143,25 +168,27 @@ export default {
       inputState: 'inputState',
       contextMenus: 'contextMenus',
       dialogs: 'dialogs',
-      appPaths: 'appPaths',
+      appPaths: 'storageData.settings.appPaths',
       detectedLocale: 'detectedLocale',
       currentDir: 'navigatorView.currentDir',
       navigatorRouteIsLoaded: 'navigatorRouteIsLoaded',
       navigationPanel: 'navigationPanel',
-      appStorageGlobalSearchData: 'appPaths.storageDirectories.appStorageGlobalSearchData',
-      globalSearchDataFiles: 'appPaths.globalSearchDataFiles',
+      appStorageGlobalSearchData: 'storageData.settings.appPaths.storageDirectories.appStorageGlobalSearchData',
       routeScrollPosition: 'routeScrollPosition',
       lastRecordedAppVersion: 'storageData.settings.lastRecordedAppVersion',
       shortcuts: 'storageData.settings.shortcuts',
       globalSearchScanDepth: 'storageData.settings.globalSearchScanDepth',
       globalSearchDisallowedPaths: 'storageData.settings.globalSearch.disallowedPaths',
       globalSearchCompressSearchData: 'storageData.settings.compressSearchData',
-      isFirstAppLaunch: 'storageData.settings.isFirstAppLaunch',
+      firstTimeActions: 'storageData.settings.firstTimeActions',
       lastSearchScanTime: 'storageData.settings.time.lastSearchScan',
       globalSearchAutoScanIntervalTime: 'storageData.settings.globalSearchAutoScanIntervalTime',
       UIZoomLevel: 'storageData.settings.UIZoomLevel',
       thumbnailStorageLimit: 'storageData.settings.thumbnailStorageLimit',
       focusMainWindowOnDriveConnected: 'storageData.settings.focusMainWindowOnDriveConnected',
+      pointerButton3: 'storageData.settings.input.pointerButtons.button3',
+      pointerButton4: 'storageData.settings.input.pointerButtons.button4',
+      animations: 'storageData.settings.animations',
       timeSinceLoadDirItems: 'navigatorView.timeSinceLoadDirItems',
       history: 'navigatorView.history',
       globalSearchScanInProgress: 'globalSearch.scanInProgress',
@@ -177,16 +204,12 @@ export default {
       storageDataSettings: 'storageData.settings',
       themeType: 'storageData.settings.theme.type',
       overlayInboundDrag: 'overlays.inboundDrag',
-      targetItems: 'contextMenus.dirItem.targetItems',
-      localServer: 'localServer',
-      fileShareIp: 'localServer.fileShare.ip',
-      fileSharePort: 'localServer.fileShare.port',
-      fileShareServer: 'localServer.fileShare.server',
-      fileShareType: 'localServer.fileShare.type',
-      fileShareItems: 'localServer.fileShare.items',
-      directoryShareIp: 'localServer.directoryShare.ip',
-      directorySharePort: 'localServer.directoryShare.port',
-      directoryShareServer: 'localServer.directoryShare.server'
+      dirItemBackground: 'storageData.settings.dirItemBackground',
+      visualFiltersApplyFiltersToMediaElements: 'storageData.settings.visualFilters.applyFiltersToMediaElements',
+      visualFiltersContrastValue: 'storageData.settings.visualFilters.contrast.value',
+      visualFiltersBrightnessValue: 'storageData.settings.visualFilters.brightness.value',
+      visualFiltersSaturationValue: 'storageData.settings.visualFilters.saturation.value',
+      windowsMainStateIsMaximized: 'windows.main.state.isMaximized',
     }),
     globalSearchScanWasInterrupted: {
       get () {
@@ -198,15 +221,47 @@ export default {
           value
         })
       }
-    }
+    },
   },
   methods: {
+    setCSSAttributes (name) {
+      if (name === 'visual-filters') {
+        let htmlNode = document.querySelector('html')
+        let invertInverse = this.themeType === 'light-filter' ? 1 : 0
+        let hueRotateInverse = this.themeType === 'light-filter' ? '180deg' : '0deg'
+        let contrastInverse = this.visualFiltersApplyFiltersToMediaElements
+          ? this.visualFiltersContrastValue
+          : 1 + (1 - this.visualFiltersContrastValue)
+        let brightnessInverse = this.visualFiltersApplyFiltersToMediaElements
+          ? this.visualFiltersBrightnessValue
+          : 1 + (1 - this.visualFiltersBrightnessValue)
+        let saturationInverse = this.visualFiltersApplyFiltersToMediaElements
+          ? this.visualFiltersSaturationValue
+          : 1 + (1 - this.visualFiltersSaturationValue)
+
+        htmlNode.style.setProperty('--visual-filter-invert', this.themeType === 'light-filter' ? 1 : 0)
+        htmlNode.style.setProperty('--visual-filter-hue-rotate', this.themeType === 'light-filter' ? '180deg' : '0deg')
+        htmlNode.style.setProperty('--visual-filter-contrast', this.visualFiltersContrastValue)
+        htmlNode.style.setProperty('--visual-filter-brightness', this.visualFiltersBrightnessValue)
+        htmlNode.style.setProperty('--visual-filter-saturation', this.visualFiltersSaturationValue)
+
+        htmlNode.style.setProperty('--visual-filter-invert-inverse', invertInverse)
+        htmlNode.style.setProperty('--visual-filter-hue-rotate-inverse', hueRotateInverse)
+        htmlNode.style.setProperty('--visual-filter-contrast-inverse', contrastInverse)
+        htmlNode.style.setProperty('--visual-filter-brightness-inverse', brightnessInverse)
+        htmlNode.style.setProperty('--visual-filter-saturation-inverse', saturationInverse)
+      }
+    },
     initEventHubListeners () {
       this.$eventHub.$on('app:method', payload => {
         this[payload.method](payload.params)
       })
     },
     initIPCListeners () {
+      electron.ipcRenderer.on('open-global-search', (event, data) => {
+        this.$store.dispatch('TOGGLE_GLOBAL_SEARCH')
+      })
+
       electron.ipcRenderer.on('store:action', (event, data) => {
         this.$store.dispatch(data.action, data.params)
       })
@@ -223,14 +278,20 @@ export default {
         this.windowBlurHandler()
       })
 
+      electron.ipcRenderer.on('window-event:maximize', (event) => {
+        this.windowsMainStateIsMaximized = true
+      })
+
+      electron.ipcRenderer.on('window-event:unmaximize', (event) => {
+        this.windowsMainStateIsMaximized = false
+      })
+
       electron.ipcRenderer.on('load:webview::failed', (event, data) => {
-        this.$eventHub.$emit('notification', {
-          action: 'update-by-type',
-          type: 'load:webview::failed',
-          closeButton: true,
-          timeout: 3000,
-          title: 'Quick view: file is not supported',
-          message: data.path.replace('file:///', '')
+        notifications.emit({
+          name: 'quickViewFileIsNotSupported',
+          props: {
+            data
+          }
         })
       })
 
@@ -238,76 +299,45 @@ export default {
         const isDone = data.receivedBytes === data.totalBytes
         data.isDone = isDone
         data.started = data.percentDone > 0
-        let actionButtons = []
         if (isDone) {
-          actionButtons = [
-            {
-              title: 'open file',
-              action: 'openDownloadedFile',
-              closesNotification: true,
-              onClick: () => {
-                this.$store.dispatch('OPEN_FILE', data.filePath)
-              }
-            },
-            {
-              title: 'show in directory',
-              action: 'showDownloadedFile',
-              closesNotification: true,
-              onClick: () => {
-                this.$store.dispatch('SHOW_DIR_ITEM_IN_DIRECTORY', { dir: data.dir, itemPath: data.filePath })
-              }
+          notifications.emit({
+            name: 'fileDownloadIsDone',
+            props: {
+              data,
+              electron,
+              store: this.$store
             }
-          ]
+          })
         }
         else {
-          actionButtons = [
-            {
-              title: 'cancel',
-              action: 'cancelDownload',
-              closesNotification: true,
-              onClick: () => {
-                electron.ipcRenderer.send('download-file:cancel', { hashID: data.hashID })
-              }
-            }
-          ]
           if (data.isPaused) {
-            actionButtons.unshift({
-              title: 'resume',
-              action: 'resumeDownloading',
-              closesNotification: false,
-              onClick: () => {
-                electron.ipcRenderer.send('download-file:resume', { hashID: data.hashID })
+            notifications.emit({
+              name: 'fileDownloadIsPaused',
+              props: {
+                data,
+                electron,
+                store: this.$store
               }
             })
           }
           else {
-            actionButtons.unshift({
-              title: 'pause',
-              action: 'pauseDownloading',
-              closesNotification: false,
-              onClick: () => {
-                electron.ipcRenderer.send('download-file:pause', { hashID: data.hashID })
+            notifications.emit({
+              name: 'fileDownloadIsInProgress',
+              props: {
+                data,
+                electron,
+                store: this.$store
               }
             })
           }
         }
-        this.$eventHub.$emit('notification', {
-          action: 'update-by-hash',
-          type: 'progress:download-file',
-          actionButtons: actionButtons,
-          closeButton: true,
-          timeout: 0,
-          title: `Downloading ${data.isUpdate ? 'update' : 'file'}`,
-          message: '',
-          progress: data,
-          hashID: data.hashID
-        })
       })
     },
     checkForAppUpdateInstalled () {
       this.$store.dispatch('CHECK_IF_UPDATE_INSTALLED')
     },
     bindKeyEvents () {
+      this.bindMouseKeyEvents()
       this.bindGeneralKeyEvents()
       this.bindGeneralMousetrapEvents()
     },
@@ -316,12 +346,12 @@ export default {
         const disallowedErrors = [
           'ResizeObserver loop limit exceeded'
         ]
-        console.log('error::App.vue::initWindowErrorHandler', event)
         if (!disallowedErrors.includes(event.message)) {
           const hashID = this.$utils.getHash()
           this.$eventHub.$emit('notification', {
             action: 'add',
             hashID,
+            colorStatus: 'red',
             timeout: 0,
             closeButton: true,
             icon: 'mdi-alert-octagon-outline',
@@ -353,20 +383,42 @@ export default {
       const shouldFocus = this.focusMainWindowOnDriveConnected
       if (previousDataExists && driveCountIncreased && shouldFocus) {
         electron.ipcRenderer.send('focus-main-app-window')
+        notifications.emit({name: 'driveWasConnected'})
       }
     },
-    extractAppBinaries () {
+    async extractAppBinaries () {
       // Moving binaries to app storage because fs.childProcess.spawn
       // and other modules cannot access it from within app.asar
       const isEnvProduction = process.env.NODE_ENV === 'production'
       if (isEnvProduction) {
-        // TODO: do not overwrite existing dirs in appStorageBin
-        // if user puts custom binaries there (need a setting switch for it)
-        require('fs-extra').copy(
-          this.appPaths.bin, 
-          this.appPaths.storageDirectories.appStorageBin
-        )
+        try {
+          await fs.promises.access(this.appPaths.bin7Zip, fs.constants.F_OK)
+          await fs.promises.access(this.appPaths.binFFMPEG, fs.constants.F_OK)
+          await fs.promises.access(this.appPaths.binYoutubeDl, fs.constants.F_OK)
+        }
+        catch (error) {
+          await fsExtra.copy(
+            this.appPaths.resourcesBin, 
+            this.appPaths.storageDirectories.appStorageBin
+          )
+          if (process.platform !== 'win32') {
+            this.getAppStorageBinDirPermissions()
+          }
+        }
       }
+    },
+    getAppStorageBinDirPermissions () {
+      const commandGetBinDirPermissionsRecursive = `chmod -R u+rwx "${this.appPaths.storageDirectories.appStorageBin}"`
+      childProcess.exec(commandGetBinDirPermissionsRecursive, (error) => {
+        if (error) {
+          notifications.emit({
+            name: 'getAppStorageBinDirPermissionsError',
+            props: {
+              error
+            }
+          })
+        }
+      })
     },
     windowBlurHandler () {
       // Force reset keyboard state to avoid pressed keys state remain true
@@ -419,30 +471,6 @@ export default {
         })
       }, 1000)
     },
-    initMediaDirectories () {
-      // TODO:
-      // - See what happens why I try to add the same image to multiple notes. 
-      // Will there be duplicates? Make sure it's not copied if it's already there
-      // - Add image gallery so that images from /notes can be reused in multiple notes
-      return new Promise((resolve, reject) => {
-        const directoriesToInit = [
-          this.appPaths.storageDirectories.appStorage,
-          this.appPaths.storageDirectories.appStorageBin,
-          this.appPaths.storageDirectories.appStorageHomeBannerMedia,
-          this.appPaths.storageDirectories.appStorageNotesMedia,
-          this.appPaths.storageDirectories.appStorageGlobalSearchData,
-          this.appPaths.storageDirectories.appStorageNavigatorThumbs
-        ]
-        const promises = []
-        directoriesToInit.forEach(path => {
-          // Attempt to create directory.
-          // Don't do anything if it's already created
-          promises.push(fs.promises.mkdir(path, { recursive: true }))
-        })
-        Promise.allSettled(promises)
-          .then(() => resolve())
-      })
-    },
     async initGlobalSearchDataScan () {
       // TODO: Check if the drive has enough space on it
       await this.initGlobalSearchDataFiles()
@@ -458,8 +486,8 @@ export default {
         const fileObject = this.appPaths.globalSearchDataFiles.find(object => object.mount === drive.mount)
         if (!fileObject) {
           this.$store.commit('PUSH', {
-            key: 'appPaths.globalSearchDataFiles',
-            value: searchDataFileObject
+            key: 'storageData.settings.appPaths.globalSearchDataFiles',
+            value: searchDataFileObject,
           })
         }
         // Check if file already exists, otherwise create it
@@ -556,7 +584,7 @@ export default {
         if (driveIsMounted) {
           task.props.driveCount++
           task.props.scans[index] = {}
-          task.props.scans[index].readStream = new Walk(
+          task.props.scans[index].readStream = new DriveWalker(
             searchDataFile.mount,
             this.globalSearchScanDepth,
             this.globalSearchDisallowedPaths
@@ -573,7 +601,6 @@ export default {
           }
           // Setup listeners
           task.props.scans[index].stream.on('error', (error) => {
-            console.log(error)
           })
           task.props.scans[index].stream.on('finish', () => {
             task.props.scannedDriveCount++
@@ -619,34 +646,14 @@ export default {
         value: Date.now()
       })
     },
-    restoreRouteScrollPosition (params) {
-      const { toRoute } = params
-      const historyItems = this.history.itemsRaw
-      const lastHistoryItem = historyItems[historyItems.length - 1]
-      const secondFromEndHistoryItem = historyItems[historyItems.length - 2]
-      const gotBackToSameNavigatorDir = toRoute.name === 'navigator' && (secondFromEndHistoryItem === this.currentDir.path)
-      const shouldRestoreScroll = toRoute.name !== 'navigator' || !gotBackToSameNavigatorDir
-      if (shouldRestoreScroll) {
-        const scrollArea = this.$utils.getContentAreaNode(toRoute.name)
-        const savedScrollPosition = this.routeScrollPosition[toRoute.name]
-        if (savedScrollPosition) {
-          scrollArea.scroll({
-            top: savedScrollPosition,
-            left: 0,
-            behavior: 'auto'
-          })
-        }
-      }
-    },
-    setRouteScrollPosition (params) {
-      const { fromRoute } = params
-      const scrollArea = this.$utils.getContentAreaNode(fromRoute.name)
-      this.routeScrollPosition[fromRoute.name] = scrollArea?.scrollTop
-    },
-    importExpress () {
-      // Lazy-import heavy module
-      if (express === undefined) { express = require('express') }
-      if (expressServer === undefined) { expressServer = express() }
+    preventHomeViewLayoutTransition () {
+      // This function prevents home banner layout shifting
+      // when transitioning from a view with an opened info panel
+      let appNode = document.querySelector('#app')
+      appNode.classList.add('layout-no-transition')
+      setTimeout(() => {
+        appNode.classList.remove('layout-no-transition')
+      }, 500)
     },
     initGlobalSearchWorker (params, workerObject) {
       return new Promise((resolve, reject) => {
@@ -656,14 +663,7 @@ export default {
           this.$eventHub.$emit('app:method', {
             method: 'initGlobalSearchDataScan'
           })
-          this.$eventHub.$emit('notification', {
-            action: 'update-by-type',
-            type: 'error:search-file-problem',
-            timeout: 8000,
-            closeButton: true,
-            title: 'Drive scan initiated',
-            message: 'One of the search files is damaged'
-          })
+          noЯtifications.emit({name: 'searchFileIsDamaged'})
           reject(new Error('One of the search files is damaged'))
         }
         else {
@@ -677,19 +677,11 @@ export default {
             }
           }
           workerObject.worker.onerror = (error) => {
-            console.log('Error in "initglobalSearchWorker"', error)
             if (error.message === 'Uncaught Error: unexpected end of file') {
               this.$eventHub.$emit('app:method', {
                 method: 'initGlobalSearchDataScan'
               })
-              this.$eventHub.$emit('notification', {
-                action: 'update-by-type',
-                type: 'error:search-file-problem',
-                timeout: 8000,
-                closeButton: true,
-                title: 'Drive scan initiated',
-                message: 'One of the search files is damaged'
-              })
+              notifications.emit({name: 'searchFileIsDamaged'})
             }
           }
           resolve()
@@ -697,16 +689,16 @@ export default {
       })
     },
     initGlobalSearchWorkerAction (params) {
-      const workerObject = this.$store.state.workers.globalSearchWorkers.find(item => item.mount === params.mount)
+      let workerObject = this.$store.state.workers.globalSearchWorkers.find(item => item.mount === params.mount)
       if (params.action === 'cancel') {
         if (workerObject) {
-          workerObject.worker.postMessage({ action: 'cancel' })
+          workerObject.worker.postMessage({action: 'cancel'})
         }
       }
       else if (params.action === 'search') {
         // If an ongoing worker exist
         if (workerObject) {
-          workerObject.worker.postMessage({ action: 'cancel' })
+          workerObject.worker.postMessage({action: 'cancel'})
           setTimeout(() => {
             // Re-init worker
             workerObject.worker = new GlobalSearchWorker()
@@ -715,13 +707,12 @@ export default {
         }
         // If worker does not exist
         else {
-          // Add worker
-          const workerObject = {
+          let newWorkerObject = {
             mount: params.mount,
             worker: new GlobalSearchWorker()
           }
-          this.$store.state.workers.globalSearchWorkers.push(workerObject)
-          this.execGlobalSearchWorkerAction(params, workerObject)
+          this.$store.state.workers.globalSearchWorkers.push(newWorkerObject)
+          this.execGlobalSearchWorkerAction(params, newWorkerObject)
         }
       }
     },
@@ -762,11 +753,11 @@ export default {
     handleFirstAppLaunch () {
       // TODO:
       // In production, some store properties resolve after this method
-      // so this.isFirstAppLaunch is always true
+      // so this.firstTimeActions.appLaunch is always true
       return
-      if (this.isFirstAppLaunch) {
+      if (this.firstTimeActions.appLaunch) {
         this.$store.dispatch('SET', {
-          key: 'storageData.settings.isFirstAppLaunch',
+          key: 'storageData.settings.firstTimeActions.appLaunch',
           value: false
         })
         this.$eventHub.$emit('notification', {
@@ -779,341 +770,6 @@ export default {
           `
         })
       }
-    },
-    stopLocalDirectoryShareServer () {
-      try {
-        this.$store.state.childProcesses.localDirectoryShareServer.kill()
-        clearInterval(this.$store.state.intervals.localDirectoryShareServerSelfDistruction)
-      }
-      catch (error) {}
-    },
-    stopLocalFileShareServer () {
-      try {
-        this.$store.state.childProcesses.localFileShareServer.kill()
-        clearInterval(this.$store.state.intervals.localFileShareServerSelfDistruction)
-      }
-      catch (error) {}
-    },
-    scheduleLocalFileShareServerSelfDistruct () {
-      try {
-        this.$store.state.childProcesses.localFileShareServer.send({
-          action: 'reset-signal::self-distruct'
-        })
-      }
-      catch (error) {}
-      clearInterval(this.$store.state.intervals.localFileShareServerSelfDistruction)
-      this.$store.state.intervals.localFileShareServerSelfDistruction = setInterval(() => {
-        try {
-          this.$store.state.childProcesses.localFileShareServer.send({
-            action: 'reset-signal::self-distruct'
-          })
-        }
-        catch (error) {}
-      }, 10000)
-    },
-    initLocalFileShareServerProcess (path) {
-      console.log('initLocalFileShareServerProcess', path)
-      this.$store.state.childProcesses.localFileShareServer = childProcess.fork('E:/Projects/Code/Apps/Desktop/Sigma file manager/Project/sigma-file-manager/src/workers/localFileShareServerWorker.js')
-      this.$store.state.childProcesses.localFileShareServer.on('message', data => {
-        console.log(data)
-      })
-      this.$store.state.childProcesses.localFileShareServer.send({
-        action: 'start-server',
-        path,
-        fileShareType: this.fileShareType,
-        fileSharePort: this.fileSharePort
-      })
-      this.scheduleLocalFileShareServerSelfDistruct()
-
-      // const subprocess = childProcess.spawn(process.argv[0], ['E:/Projects/Code/Apps/Desktop/Sigma file manager/Project/sigma-file-manager/src/workers/localFileShareServerWorker.js'], {
-      //   // detached: true,
-      //   // stdio: 'ignore'
-      // })
-
-      // let subprocess = childProcess.exec(
-      //   payload.command,
-      //   (error, stdout, stderr) => {
-      //     if (error) {
-      //       progress.eta = 0
-      //       notificationData.timeout = 5000
-      //       notificationData.title = status.isCanceled ? 'Download canceled' : 'Download failed'
-      //       notificationData.actionButtons = []
-      //       eventHub.$emit('notification', notificationData)
-      //       return
-      //     }
-      //   }
-      // )
-
-      // this.$store.state.childProcesses.localFileShareServer.stdout.on('data', (data) => {
-      //   console.log('express', data)
-      //   // const dataArray = data.split(' ')
-      //   //   .filter(item => item !== '')
-      //   // // Get only info from 'download' phase
-      //   // if (dataArray[1].endsWith('%') && dataArray[5] !== 'unknown') {
-      //   //   // Update progess data
-      //   //   progress.started = true
-      //   //   progress.percentDone = dataArray[1]?.replace(/\%/, '')
-      //   //   progress.speed = dataArray[5]
-      //   //   progress.eta = dataArray[7]?.replace(/\[download\]/, '')
-      //   //   if (!status.isCanceled) {
-      //   //     // Update notification
-      //   //     const actionButtons = [
-      //   //       {
-      //   //         title: 'cancel',
-      //   //         action: '',
-      //   //         closesNotification: true,
-      //   //         onClick: () => {
-      //   //           status.isCanceled = true
-      //   //           dispatch('TERMINATE_PROCESS', mainExecProcess)
-      //   //         }
-      //   //       }
-      //   //     ]
-      //   //     notificationData.message = ''
-      //   //     notificationData.actionButtons = actionButtons
-      //   //     eventHub.$emit('notification', notificationData)
-      //   //   }
-      //   // }
-      // })
-    },
-    async initLocalDirectoryShare () {
-      await this.stopLocalDirectoryShareServer()
-      this.dialogs.localShareManagerDialog.value = true
-      this.dialogs.localShareManagerDialog.data.shareType = 'directory'
-      this.localServer.directoryShare.item = this.selectedDirItems.getLast()
-      this.directoryShareIp = await this.getLocalIPv4()
-      this.directorySharePort = await getPort({ port: getPort.makeRange(55000, 55999) })
-      this.localServer.directoryShare.address = `${this.directoryShareIp}:${this.directorySharePort}/ftp`
-      const path = this.selectedDirItems.getLast().path
-      this.serveDirectoryLocally(path)
-      this.$eventHub.$emit('notification', {
-        action: 'update-by-type',
-        type: 'directoryShare',
-        actionButtons: [
-          {
-            title: 'stop server',
-            onClick: () => {
-              this.stopLocalDirectoryShareServer()
-              this.dialogs.localShareManagerDialog.value = false
-            },
-            closesNotification: true
-          },
-          {
-            title: 'copy address',
-            onClick: () => {
-              this.$store.dispatch(
-                'COPY_TEXT_TO_CLIPBOARD',
-                {text: `${this.directoryShareIp}:${this.directorySharePort}/ftp`}
-              )
-            },
-            closesNotification: false
-          }
-        ],
-        closeButton: false,
-        timeout: 0,
-        title: 'Directory is now accessible from local devices',
-        message: `Address: ${this.directoryShareIp}:${this.directorySharePort}/ftp`
-      })
-    },
-    async initLocalFileShare () {
-      await this.stopLocalFileShareServer()
-      this.dialogs.localShareManagerDialog.value = true
-      this.dialogs.localShareManagerDialog.data.shareType = 'file'
-      this.localServer.fileShare.item = this.targetItems.getLast()
-      this.fileShareIp = await this.getLocalIPv4()
-      this.fileSharePort = await getPort({ port: getPort.makeRange(56000, 56999) })
-      this.localServer.fileShare.address = `${this.fileShareIp}:${this.fileSharePort}`
-      const path = this.targetItems.getLast().path
-      this.serveFilesLocally(path)
-      this.$eventHub.$emit('notification', {
-        action: 'update-by-type',
-        type: 'fileShare',
-        actionButtons: [
-          {
-            title: 'stop server',
-            onClick: () => {
-              this.stopLocalFileShareServer()
-              this.dialogs.localShareManagerDialog.value = false
-            },
-            closesNotification: true
-          },
-          {
-            title: 'copy address',
-            onClick: () => {
-              this.$store.dispatch(
-                'COPY_TEXT_TO_CLIPBOARD',
-                {text: `${this.fileShareIp}:${this.fileSharePort}`}
-              )
-            },
-            closesNotification: false
-          }
-        ],
-        closeButton: false,
-        timeout: 0,
-        title: 'Local file share is active',
-        message: `Address: ${this.fileShareIp}:${this.fileSharePort}`
-      })
-    },
-    getLocalIPv4 () {
-      // TODO:
-      // - When offline, will throw:
-      //   "Error: getaddrinfo ENOTFOUND www.google.com"
-      //   And local file share will not be accessible.
-      //   To fix use another method for detecting ipv4 address.
-      return new Promise((resolve, reject) => {
-        const socket = net.createConnection(80, 'www.google.com')
-        socket.on('connect', () => {
-          resolve(socket.address().address)
-          socket.end()
-        })
-        socket.on('error', (error) => {
-          reject(error)
-        })
-      })
-    },
-    serveDirectoryLocally (path) {
-      this.generateQrCode({
-        container: document.querySelector('#qr-code'),
-        content: this.localServer.directoryShare.address
-      })
-      this.importExpress()
-      this.initFTPserver(path)
-      this.initFTPclientFileUpload(path)
-      this.directoryShareServer = expressServer.listen(this.directorySharePort)
-    },
-    serveFilesLocally (path) {
-      this.generateQrCode({
-        container: document.querySelector('#qr-code'),
-        content: `${this.fileShareIp}:${this.fileSharePort}`
-      })
-      this.importExpress()
-      // this.serveFile(path)
-      this.initLocalFileShareServerProcess(path)
-    },
-    // serveFile (path) {
-    //   expressServer.get('/', (request, response, next) => {
-    //     // Stream (display) file
-    //     if (this.fileShareType === 'stream') {
-    //       response.sendFile(path)
-    //     }
-    //     // Push download
-    //     else if (this.fileShareType === 'download') {
-    //       response.download(path, (error) => {
-    //         if (error) { throw error }
-    //       })
-    //     }
-    //   })
-    //   this.fileShareServer = expressServer.listen(this.fileSharePort)
-    // },
-    async initFTPserver (path) {
-      // TODO:
-      // Use https://www.npmjs.com/package/express-state
-      // To pass dirItems data and generate dirItem nodes freely,
-      // instead of using built-in {files} literal provided by serve-index
-      // Or use Vue SSR https://ssr.vuejs.org/guide/#using-a-page-template
-      // And send rendered on the server side HTML to the client side.
-      // const processedHTML = fs.readFileSync(PATH.join(__static, 'server', '/ftp.html'))
-      //  .replace('EXPRESS_REPLACE_SERVER_DATA', JSON.stringify(dirItemsData))
-      expressServer.use(express.static(PATH.join(__static, 'server', 'public')))
-      expressServer.use(
-        '/ftp',
-        express.static(path, {
-          setHeaders: (res, path, stat) => {
-            res.set('Content-Disposition', 'inline')
-          }
-        }),
-        serveIndex(path, {
-          icons: true,
-          template: PATH.join(__static, 'server', '/ftp.html')
-          // template: processedHTML
-        })
-      )
-    },
-    initFTPclientFileUpload (path) {
-      expressServer.use(fileUpload({
-        useTempFiles: true,
-        tempFileDir: path
-      }))
-
-      expressServer.post('/uploaded', (req, res) => {
-        if (!req.files || Object.keys(req.files).length === 0) {
-          return res.status(400).send('No files were uploaded')
-        }
-        // If only 1 file is uploaded, put it into an array
-        const uploadedItems = Array.isArray(req.files.items)
-          ? req.files.items
-          : [req.files.items]
-
-        // Write each uploaded file to storage
-        uploadedItems.forEach(item => {
-          // Get available names for each item
-          item.path = this.$utils.getUniquePath(PATH.join(path, item.name))
-          // Write files to storage
-          item.mv(item.path, (error) => {
-            if (error) {
-              return res.status(500).send(error)
-            }
-            else {
-              res.send(
-                `
-                  <style>
-                    body {
-                      margin: 0
-                    }
-                    
-                    button {
-                      cursor: pointer;
-                      padding: 8px 18px;
-                      margin-top: 12px;
-                      font-size: 12px;
-                      font-weight: 600;
-                      letter-spacing: 1px;
-                      border-radius: 4px;
-                      text-transform: uppercase;
-                      color: #bdbdbd;
-                      border: 2px solid #bdbdbd;
-                      background: transparent;
-                    }
-
-                    .container {
-                      display: flex;
-                      flex-direction: column;
-                      align-items: center;
-                      justify-content: center;
-                      height: 100%;
-                      width: 100%;
-                      background-color: #37474f;
-                      color: #bdbdbd;
-                      font-size: 36px;
-                    }
-                  </style>
-                  <div class="container">
-                    <div>Files were uploaded</div>
-                    <button onclick="window.location.pathname = '/ftp'">Go back</button>
-                  </div>
-                `
-              )
-            }
-          })
-        })
-      })
-    },
-    generateQrCode (options) {
-      const defaultOptions = {
-        errorCorrectionLevel: 'H',
-        margin: 2,
-        width: 96,
-        height: 96
-      }
-      options = { ...defaultOptions, ...options }
-      qrCode.toCanvas(options.content, options, (error, canvas) => {
-        if (error) { throw error }
-        // Remove previous qr element
-        while (options.container.firstChild) {
-          options.container.removeChild(options.container.firstChild)
-        }
-        // Create qr element
-        options.container.appendChild(canvas)
-      })
     },
     openWithExternalProgram (app) {
       this.$store.commit('OPEN_WITH_CUSTOM_APP', app)
@@ -1174,13 +830,17 @@ export default {
             // Merge storage value with store value if it's an object,
             // otherwise overwrite the value in store.
             // This will prevent errors, when new object properties are added with updates
+            let updatedValue = {}
             if (this.$utils.getDataType(value) === 'object' && this.$utils.getDataType(propertyInStore) === 'object') {
-              value = Object.assign({}, propertyInStore, value)
+              updatedValue = Object.assign({}, propertyInStore, value)
+            }
+            else {
+              updatedValue = value
             }
             // Update store. Write updated settings back to the storage file
             await this.$store.dispatch('SET', {
               key,
-              value,
+              value: updatedValue,
               options: {
                 updateStorage: payload.fileName === 'settings.json'
               }
@@ -1217,7 +877,6 @@ export default {
       //     data[key] = this.$sharedUtils.shortcutRawToReadable(data[key])
       //   })
       // }
-      // console.log('formatAppStorageData', data)
       return data
     },
     removeOutdatedDirItems (items) {
@@ -1294,7 +953,7 @@ export default {
             fill: 'forwards'
           }
         )
-        this.animateHomeBanner()
+        this.animateHomeBannerIn()
       }
     },
     removeLoadingScreen () {
@@ -1310,35 +969,32 @@ export default {
         }, fadeOutTimeout)
       }
     },
-    animateHomeBanner (options = {}) {
-      setTimeout(() => {
-        animate()
-      }, options.delay ? 400 : 0)
-
-      function animate () {
-        try {
-          const homeBannerImgNode = document.querySelector('.media-banner img')
-          const homeBannerVideoNode = document.querySelector('.media-banner video')
-          let targetNode
-          if (homeBannerImgNode) {
-            targetNode = homeBannerImgNode
-          }
-          else if (homeBannerVideoNode) {
-            targetNode = homeBannerVideoNode
-          }
-          targetNode.animate(
-            [
-              { transform: 'scale(1.1)' },
-              { transform: 'scale(1)' }
-            ],
-            {
-              easing: 'cubic-bezier(.07,1.04,.74,1)',
-              duration: 2000,
-              fill: 'forwards'
-            }
-          )
+    animateHomeBanner (params) {
+      try {
+        const homeBannerImgNode = document.querySelector('.media-banner img')
+        const homeBannerVideoNode = document.querySelector('.media-banner video')
+        let targetNode
+        if (homeBannerImgNode) {
+          targetNode = homeBannerImgNode
         }
-        catch (error) {}
+        else if (homeBannerVideoNode) {
+          targetNode = homeBannerVideoNode
+        }
+        targetNode.style.transform = params.transform
+      }
+      catch (error) {}
+    },
+    animateHomeBannerIn () {
+      this.$nextTick(() => {
+        this.animateHomeBanner({transform: 'scale(1)'})
+      })
+    },
+    animateHomeBannerOut () {
+      if (this.animations.onRouteChangeMediaBannerIn) {
+        this.animateHomeBanner({transform: 'scale(1.2)'})
+      }
+      else {
+        this.animateHomeBanner({transform: 'scale(1)'})
       }
     },
     initGlobalShortcuts () {
@@ -1362,8 +1018,6 @@ export default {
         events: ['mousemove', 'keydown', 'mousedown', 'touchstart'],
         onIdle: () => this.handleAppIdleState(),
         onActive: () => this.handleAppActiveState(),
-        onHide: () => {},
-        onShow: () => {},
         keepTracking: true,
         startAtIdle: false
       }).start()
@@ -1383,49 +1037,40 @@ export default {
       this.appStatus.state = 'active'
       this.initStorageDevicesWatcher()
     },
-    handleThumbCacheRemoval () {
-      this.getDirItemTotalSize([this.appPaths.storageDirectories.appStorageNavigatorThumbs])
-        .then((thumbDirSizeInBytes) => {
-          const sizeLimitInBytes = this.thumbnailStorageLimit * 1000 * 1000
-          if (thumbDirSizeInBytes > sizeLimitInBytes) {
-            this.$store.dispatch('FETCH_DIR_ITEM_INFO', this.appPaths.storageDirectories.appStorageNavigatorThumbs)
-              .then((item) => {
-                this.$store.dispatch('DELETE_DIR_ITEMS', {
-                  items: [item],
-                  options: {
-                    skipSafeCheck: true,
-                    silent: true
-                  }
-                })
-              })
-          }
-        })
-        .catch((error) => {
-          console.log('error::App.vue::handleThumbCacheRemoval', error)
-        })
-    },
-    async getDirItemTotalSize (paths) {
-      let totalSize = 0
-      for (let index = 0; index < paths.length; index++) {
-        const path = paths[index]
-        const pathStat = fs.statSync(path)
-        if (pathStat.isFile()) {
-          totalSize += pathStat.size
-        }
-        else {
-          const dirSize = await this.getDirectorySize(path)
-          totalSize += dirSize
+    async handleThumbCacheRemoval () {
+      let appStorageNavigatorThumbsDirPath = this.appPaths.storageDirectories.appStorageNavigatorThumbs
+      try {
+        let appThumbDirSizeInBytes = await fsInfo.getDirItemTotalSize([appStorageNavigatorThumbsDirPath])
+        let appThumbDirSizeLimitInBytes = this.thumbnailStorageLimit * 1024 * 1024
+        let appThumbDirSizeLimitReadable = this.$utils.prettyBytes(appThumbDirSizeLimitInBytes)
+        let appThumbDirSizeLimitExceeded = appThumbDirSizeInBytes > appThumbDirSizeLimitInBytes
+        if (appThumbDirSizeLimitExceeded) {
+          let dirent = await this.$store.dispatch('GET_DIR_ITEM_INFO', appStorageNavigatorThumbsDirPath)
+          await this.$store.dispatch('DELETE_DIR_ITEMS', {
+            items: [dirent],
+            options: {
+              skipSafeCheck: true,
+              silent: true
+            }
+          })
+          notifications.emit({
+            name: 'removeAppThumbsDirSuccess',
+            props: {
+              thumbDirSizeLimit: appThumbDirSizeLimitReadable
+            }
+          })
         }
       }
-      return totalSize
-    },
-    async getDirectorySize (path) {
-      return new Promise((resolve, reject) => {
-        getDirSize(path, { type: 'raw', stopOnError: false }, (error, size) => {
-          if (error) { reject(error) }
-          resolve(size)
+      catch (error) {
+        if (error.code === 'ENOENT') {return}
+        notifications.emit({
+          name: 'removeAppThumbsDirError',
+          props: {
+            thumbDirPath: appStorageNavigatorThumbsDirPath,
+            error
+          }
         })
-      })
+      }
     },
     initGlobalSearchDataWatcher () {
       if (process.env.NODE_ENV !== 'development') {
@@ -1454,6 +1099,23 @@ export default {
       }
       catch (error) {
         throw Error(error)
+      }
+    },
+    bindMouseKeyEvents () {
+      window.addEventListener('mouseup', this.mouseupHandler)
+    },
+    mouseupHandler (mouseupEvent) {
+      if (mouseupEvent.button === 3) {
+        if (this.pointerButton3.onMouseUpEvent.action !== 'default') {
+          mouseupEvent.preventDefault()
+          this.$store.dispatch(this.pointerButton3.onMouseUpEvent.action)
+        }
+      }
+      else if (mouseupEvent.button === 4) {
+        if (this.pointerButton4.onMouseUpEvent.action !== 'default') {
+          mouseupEvent.preventDefault()
+          this.$store.dispatch(this.pointerButton4.onMouseUpEvent.action)
+        }
       }
     },
     bindGeneralKeyEvents () {
@@ -1505,16 +1167,21 @@ export default {
               .map((_, index) => `${shortcutStaticPart}${index + 1}`)
             mousetrap.bind(routeShortcuts, (event) => {
               const index = parseInt(event.code.replace('Digit', ''))
-              value.action.props = this.navigationPanel.items[index - 1]
+              value.action.options = this.navigationPanel.items[index - 1]
               this.$store.dispatch('SHORTCUT_ACTION', { event, value })
             }, 'keydown')
           }
           else {
             // Bind specified action to shortcuts
             if (!value.isReadOnly) {
-              mousetrap.bind(value.shortcut, (event) => {
+              // Add additional (duplicate) listeners
+              let shortcut = value.shortcut
+              if (value.shortcut === 'ctrl++') {
+                shortcut = ['ctrl++', 'ctrl+=']
+              }
+              mousetrap.bind(shortcut, (event) => {
                 this.$store.dispatch('SHORTCUT_ACTION', { event, value })
-              }, 'keydown')
+              }, value.eventName ?? 'keydown')
             }
           }
         }
@@ -1525,144 +1192,60 @@ export default {
 </script>
 
 <style>
-@font-face {
-  font-family: 'Roboto-Light';
-  src: url('./assets/fonts/Roboto/Roboto-Light.ttf') format('truetype')
+@import url('./styles/globalFonts.css');
+@import url('./styles/globalVariables.css');
+@import url('./styles/globalVuetifyOverrides.css');
+@import url('./styles/globalHelpers.css');
+@import url('./styles/globalTransitions.css');
+@import url('./styles/scrollbars.css');
+
+html {
+  overflow: hidden !important;
+  filter:
+    invert(var(--visual-filter-invert)) 
+    hue-rotate(var(--visual-filter-hue-rotate)) 
+    contrast(var(--visual-filter-contrast)) 
+    brightness(var(--visual-filter-brightness))
+    saturate(var(--visual-filter-saturation));
 }
 
-@font-face {
-  font-family: 'Roboto-Regular';
-  src: url('./assets/fonts/Roboto/Roboto-Regular.ttf') format('truetype')
+img,
+picture,
+video,
+.media-banner__inner__container--left,
+.overlay--window-transparency-effect__media,
+#loading-screen__container,
+#app[route-name='home'] 
+  .window-toolbar__item {
+    filter: 
+      invert(var(--visual-filter-invert-inverse)) 
+      hue-rotate(var(--visual-filter-hue-rotate-inverse)) 
+      contrast(var(--visual-filter-contrast-inverse)) 
+      brightness(var(--visual-filter-brightness-inverse))
+      saturate(var(--visual-filter-saturation-inverse));
+  }
+
+.media-banner__media-item-glow {
+  filter: 
+    invert(var(--visual-filter-invert-inverse)) 
+    hue-rotate(var(--visual-filter-hue-rotate-inverse)) 
+    contrast(var(--visual-filter-contrast-inverse)) 
+    brightness(var(--visual-filter-brightness-inverse))
+    saturate(var(--visual-filter-saturation-inverse))
+    blur(32px);
 }
 
-@font-face {
-  font-family: 'Roboto-Bold';
-  src: url('./assets/fonts/Roboto/Roboto-Bold.ttf') format('truetype')
-}
-
-#app {
-  --nav-panel-width: 64px;
-  --nav-panel-active-indicator-width: 4px;
-  --nav-panel-icon-width: calc(
-    var(--nav-panel-width) -
-    (var(--nav-panel-active-indicator-width) * 2)
-  );
-  --scrollbar-width: 14px;
-  --window-toolbar-height: 32px;
-  --action-toolbar-height: 42px;
-  --workspace-area-toolbar-height: 48px;
-  --header-height: calc(
-    var(--window-toolbar-height) +
-    var(--action-toolbar-height)
-  );
-}
-
-#app[data-theme-type="light"] {}
-
-#app[data-theme-type="dark"] {
-  --main-color-1: rgba(47, 52, 68);
-  --key-color-1: #607d8b;
-  --key-color-1-darker-1: rgb(55, 68, 88);
-  --highlight-color-1: rgba(255, 255, 255, 0.3);
-  --highlight-color-2: rgba(255, 255, 255, 0.2);
-  --highlight-color-3: rgba(255, 255, 255, 0.1);
-  --highlight-color-4: rgba(255, 255, 255, 0.05);
-  --highlight-color-5: rgba(255, 255, 255, 0.01);
-  --bg-color-1-value: 36, 38, 44;
-  --bg-color-2-value: 29, 31, 37;
-  --bg-color-3-value: 47, 52, 68;
-  --bg-color-1: rgb(var(--bg-color-1-value));
-  --bg-color-2: rgb(var(--bg-color-2-value));
-  --bg-color-3: rgb(var(--bg-color-3-value));
-  --color-1: #fafafa;
-  --color-2: #f5f5f5;
-  --color-3: #eeeeee;
-  --color-4: #e0e0e0;
-  --color-5: #bdbdbd;
-  --color-6: #9e9e9e;
-  --color-7: #757575;
-  --color-8: #616161;
-  --color-1-alt-1: rgb(96, 125, 139, 0.5);
-  --code-bg-color: rgb(0, 0, 0, 0.1);
-  --table-color: #616161;
-  --icon-color-1: #546e7a;
-  --icon-color-2: #9e9e9e;
-  --icon-color-3: #607d8b;
-  --icon_active-color-1: rgb(96, 125, 139, 0.2);
-  --title-color-1: #9e9e9e;
-  --sub-title-color-1: #757575;
-  --sub-title-color-2: #9e9e9e;
-  --button-bg-color-1: rgb(255, 255, 255, 0.05);
-  --button-color-1: rgb(255, 255, 255, 0.4);
-  --button-bg-color-2: rgb(96, 125, 139, 0.4);
-  --button-color-2: rgb(255, 255, 255, 0.7);
-  --radio-bg-color-1: rgb(255, 255, 255, 0.4);
-  --radio-text-color-1: rgb(255, 255, 255, 0.4);
-  --input-bg-color-1: rgb(255, 255, 255, 0.4);
-  --input_active-track-color-1: var(--key-color-1-darker-1);
-  --input_active-thumb-color-1: var(--key-color-1);
-  --input-track-color-1: rgba(24, 26, 32);
-  --input-thumb-color-1: var(--color-6);
-  --divider-color-1: #202020;
-  --divider-color-2: #333;
-  --shadow-1: -6px 8px 24px rgb(0, 0, 0, 0.2),
-              -1px 1px 4px rgb(0,0,0,0.05);
-  --shadow-2: 6px 8px 24px rgb(0, 0, 0, 0.1),
-              1px 1px 4px rgb(0,0,0,0.05);
-  --shadow-2__hover: 8px 12px 36px rgb(0, 0, 0, 0.2),
-              1px 1px 4px rgb(0,0,0,0.05);
-  --shadow-x3: -6px 8px 24px rgb(0, 0, 0, 0.2),
-              -1px 1px 4px rgb(0,0,0,0.05);
-  --shadow-x4: -6px 8px 24px rgb(0, 0, 0, 0.2),
-              -1px 1px 4px rgb(0,0,0,0.05);
-  --shadow-x4_hover: -16px 20px 48px rgb(0,0,0,0.3);
-  --scrollbar-color: rgb(56, 58, 64, 0.5);
-  --app-content-bg-color: var(--bg-color-2);
-  --info-panel-preview-container-bg-color: #37474f;
-  --info-panel-preview-container-color: #607d8b;
-  --nav-panel-color-1: #9e9e9e;
-  --nav-panel-color_hover: rgb(120, 144, 156, 0.15);
-  --nav-panel-icon-color-1: #78909c;
-  --nav-panel-indicator-color: #455a64;
-  --nav-panel-item-bg-color_active:rgb(120, 144, 156, 0.05);
-  --info-panel-bg-color: var(--bg-color-2);
-  --info-panel-shadow: 0px 4px 48px rgb(0,0,0,0.05),
-                      0px 4px 2px rgb(0,0,0,0.3);
-  --info-panel-title-bg-color: #bdbdbd;
-  --note-color-bar-opacity: 0.5;
-  --context-menu-shadow: -4px 8px 12px rgb(0,0,0,0.05),
-                        -10px 20px 76px rgb(0,0,0,0.4);
-  --context-menu-bg-color: var(--bg-color-1);
-  --context-menu-bg-color_hover: rgb(255, 255, 255, 0.03);
-  --list-menu-bg-color_active: rgb(69, 90, 100, 0.2);
-  --dir-item-card-bg: transparent;
-  --dir-item-card-border: #111;
-  --dir-item-card-border-2: rgb(255, 255, 255, 0.4);
-  --dir-item-card-border-3: rgb(255, 255, 255, 0.08);
-  --dir-item-card-checkbox-color: rgb(159, 168, 218, 0.6);
-  --dir-item-card-icon-color: #9e9e9e;
-  --toggle-btn-bg-color: #333;
-  --list-item-mask-color: rgb(255, 255, 255, 0.6);
-  --blockquote-bg: rgb(232, 234, 246, 0.05);
-  --blockquote-border: rgba(56, 124, 170, 0.384);
-}
-
-/* STYLES */
 #app,
 .v-application,
 .theme--light.v-card {
   color: var(--color-6) !important;
-  font-family: 'Roboto-Regular',
-                /* Fix for: some emoji not
-                being displayed properly */
-                'Segoe UI Emoji',
-                'Apple Color Emoji',
-                'Droid Sans Fallback',
-                'Noto Color Emoji';
-}
-
-*:not(.v-icon) {
-  /* letter-spacing: 0.5px; */
+  font-family: 
+    'Roboto-Regular',
+    /* Fix for: some emoji not being displayed properly */
+    'Segoe UI Emoji',
+    'Apple Color Emoji',
+    'Droid Sans Fallback',
+    'Noto Color Emoji';
 }
 
 code,
@@ -1671,150 +1254,19 @@ pre {
   font-size: 18px;
 }
 
-html {
-  overflow: hidden !important;
+i {
+  font-style: unset;
 }
 
 .app-content {
-  background-color: var(--bg-color-1);
   padding-top: var(--header-height) !important;
+  background-color: var(--app-content-bg-color);
 }
 
-.app-content .v-icon {
-  color: var(--icon-color-2);
-}
-
-/* Helpers */
-.fade-mask--top-and-bottom {
-  mask: linear-gradient(transparent, black 10%, black 90%, transparent 100%);
-}
-
-.fade-mask--top {
-  mask: linear-gradient(to bottom, transparent 0%, black var(--fade-mask-top));
-}
-
-.fade-mask--bottom {
-  mask: linear-gradient(to top, transparent 0%, black var(--fade-mask-bottom));
-}
-
-.fade-mask--bottom--10 {
-  mask-image: linear-gradient(to top, transparent 0%, black 10%);
-}
-
-.fade-mask--bottom--15 {
-  mask-image: linear-gradient(to top, transparent 0%, black 15%);
-}
-
-.fade-mask--bottom--25 {
-  mask-image: linear-gradient(to top, transparent 0%, black 25%);
-}
-
-.fade-mask--bottom--35 {
-  mask-image: linear-gradient(to top, transparent 0%, black 35%);
-}
-
-.cursor-pointer {
-  cursor: pointer;
-  user-select: none;
-}
-
-.unselectable {
-  user-select: none;
-}
-
-.opacity-1 {
-  opacity: 1;
-}
-
-.opacity-0 {
-  opacity: 0;
-}
-
-.is-hidden {
-  display: none !important;
-}
-
-/* VUETIFY OVERWRITES */
-.v-navigation-drawer,
-.window-toolbar__header-container,
-.v-toolbar,
-.v-footer,
-.v-main {
-  transition: 0.5s ease !important;
-}
-
-button.v-btn.v-btn--disabled
-  .v-btn__content,
-button.v-btn.v-btn--disabled
-  .v-btn__content
-    i.v-icon {
-      color: var(--color-7) !important;
-    }
-
-.v-btn.v-btn--icon {
-  color: var(--color-6) !important;
-}
-
-.v-badge__badge[size="16"] {
-  border-radius: 16px !important;
-  font-size: 12px !important;
-  height: 16px !important;
-  width: 16px !important;
-  min-width: 16px !important;
-  padding-top: 0px !important;
-}
-
-.v-list {
-  color: var(--color-6) !important;
-}
-
-.v-list-item.v-list-item--active {
-  color: var(--color-5) !important;
-  background-color: var(--highlight-color-6) !important;
-}
-
-.v-list-item:not(.v-list-item--active):not(.v-list-item--disabled) {
-  color: var(--color-5) !important;
-}
-
-.v-list-item__mask {
-  background: var(--list-item-mask-color) !important;
-  padding: 4px 2px;
-}
-
-.v-messages {
-  color: var(--color-7) !important;
-}
-
-.v-badge__badge[size="16"] .v-icon {
-  height: 0 !important;
-}
-
-.inline-code--light {
-  border-radius: 4px;
-  background: rgb(255, 255, 255, 0.1);
-  padding: 3px 6px;
-}
-
-.code-block {
-  width: 100%;
-  overflow-y: auto;
-  padding: 8px 12px;
-  background: rgba(0, 0, 0, 0.2);
-  border-radius: 8px;
-}
-
-*[shadow~="left"][shadow~="x4"] {
-  box-shadow: var(--shadow-x4);
-}
-
-*[shadow~="left"][shadow~="x4"]:hover {
-  box-shadow: var(--shadow-x4_hover);
-}
-
-*[shadow="x3"] {
-  box-shadow: var(--shadow-x3) !important;
-}
+.app-content 
+  .v-icon {
+    color: var(--icon-color-2);
+  }
 
 .v-overlay--active,
 .v-overlay__scrim {
@@ -1827,7 +1279,6 @@ button.v-btn.v-btn--disabled
   }
 
 .v-tooltip__content {
-  opacity: 1 !important;
   color: var(--color-5) !important;
   background: rgba(41, 43, 51, 0.95) !important;
   box-shadow: 0 8px 32px rgb(0,0,0,0.5);
@@ -1844,98 +1295,25 @@ button.v-btn.v-btn--disabled
   font-size: 15px;
 }
 
+.tooltip__shortcut-list-item {
+  margin-bottom: 8px;
+}
+
+.tooltip__modifier-list-item {
+  margin-bottom: 8px;
+}
+
+.tooltip__modifier-list__title {
+  margin-top: 8px;
+  font-size: 12px;
+  text-transform: uppercase;
+}
+
 .tooltip__shortcut {
   font-size: 14px;
   color: var(--color-6);
 }
 
-/* Note: this method doesn't break alignment inside grid */
-.truncate-line {
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-/* Note: this method breaks alignment inside grid */
-.line-clamp-1 {
-  display: -webkit-box;
-  -webkit-box-orient: vertical;
-  -webkit-line-clamp: 1;
-  overflow: hidden;
-}
-
-.line-clamp-2 {
-  display: -webkit-box;
-  -webkit-box-orient: vertical;
-  -webkit-line-clamp: 2;
-  overflow: hidden;
-}
-
-.line-clamp-3 {
-  display: -webkit-box;
-  -webkit-box-orient: vertical;
-  -webkit-line-clamp: 3;
-  overflow: hidden;
-}
-
-/* Scrollbars */
-body::-webkit-scrollbar,
-html::-webkit-scrollbar,
-.custom-scrollbar::-webkit-scrollbar {
-  display: none !important;
-}
-
-.custom-scrollbar {
-  overflow: overlay !important;
-}
-
-.custom-scrollbar:hover::-webkit-scrollbar {
-  width: 16px;
-  display: block !important;
-}
-
-#home-route .custom-scrollbar:hover::-webkit-scrollbar {
-  display: none !important;
-}
-
-.custom-scrollbar::-webkit-scrollbar-thumb:vertical,
-.custom-scrollbar::-webkit-scrollbar-thumb:horizontal {
-  background-color: var(--scrollbar-color);
-  overflow: overlay;
-  min-height: 50px;
-  background-clip: padding-box;
-  border: 6px solid transparent;
-}
-
-.custom-scrollbar::-webkit-scrollbar-track:vertical,
-.custom-scrollbar::-webkit-scrollbar-track:horizontal {
-  background-color: transparent;
-}
-
-.custom-scrollbar::-webkit-scrollbar-thumb:vertical,
-.custom-scrollbar::-webkit-scrollbar-thumb:horizontal {
-  background-color: transparent;
-}
-
-.custom-scrollbar:hover::-webkit-scrollbar-thumb:vertical,
-.custom-scrollbar:hover::-webkit-scrollbar-thumb:horizontal {
-  background-color: var(--scrollbar-color);
-  /* border: 2px solid transparent */
-}
-
-.custom-scrollbar::-webkit-scrollbar-thumb:vertical:hover,
-.custom-scrollbar::-webkit-scrollbar-thumb:horizontal:hover {
-  background-color: var(--scrollbar-color);
-  border: 1px solid transparent
-}
-
-.custom-scrollbar::-webkit-scrollbar-thumb:vertical:active,
-.custom-scrollbar::-webkit-scrollbar-thumb:horizontal:active {
-  background-color: var(--scrollbar-color);
-  border: 1px solid transparent
-}
-
-/* Override scome scrollbar styles */
 .v-list-item__action:last-of-type:not(:only-child) {
   margin-left: 4px !important;
 }
@@ -1952,7 +1330,6 @@ html::-webkit-scrollbar,
   z-index: 0 !important;
 }
 
-/* General list menues */
 .list-menu__title {
   display: flex;
   justify-content: space-between;
@@ -1960,101 +1337,8 @@ html::-webkit-scrollbar,
   padding: 8px 8px;
 }
 
-/* Menu & menu > list */
 .context-menu__container {
   background: var(--context-menu-bg-color) !important;
-}
-
-.v-menu__content {
-  box-shadow: var(--context-menu-shadow) !important;
-  z-index: 10;
-}
-
-.v-menu__content .v-list {
-  background: var(--context-menu-bg-color) !important;
-  border-radius: 0 !important;
-}
-
-.v-list-item__action:first-child,
-.v-list-item__icon:first-child {
-  margin-right: 16px !important;
-}
-
-.v-list-item__title {
-  font-size: 14px !important;
-  color: var(--color-6) !important;
-}
-
-.v-list-item .v-icon {
-  color: var(--color-6) !important;
-}
-
-.v-menu__content .v-list-item:not(.inactive) .v-list-item__content,
-.v-menu__content .v-list-item:not(.inactive) .v-list-item__title,
-.v-menu__content .v-list-item:not(.inactive) .v-icon {
-  color: var(--color-6) !important;
-}
-
-.v-menu__content .v-list-item:not(.inactive):hover {
-  background-color: var(--context-menu-bg-color_hover);
-}
-
-.v-menu__content .v-divider {
-  border-color: var(--divider-color-1) !important;
-}
-
-.v-menu__content .v-list-item[is-active='true'] {
-  background-color: var(--list-menu-bg-color_active) !important;
-}
-
-/* .v-menu__content .v-list-item .v-list-item__content {
-  color: var(--color-5) !important;
-} */
-
-.v-list-item .v-list-item__subtitle {
-  color: var(--color-7) !important;
-}
-
-.v-input--radio-group--column .v-input--radio-group__input {
-  /* FIX: tooltips are centered to the container because input items
-    stretch to the whole width of the container */
-  align-items: flex-start;
-}
-
-/* Misc components */
-.divider-color-1 {
-  border-color: var(--divider-color-1) !important;
-}
-
-.divider-color-2 {
-  border-color: var(--divider-color-2) !important;
-}
-
-/* Cards */
-.v-card {
-  background-color: var(--bg-color-1) !important
-}
-
-/* Nav panel */
-.nav-panel__main-content {
-  overflow-y: overlay !important;
-  overflow-x: hidden !important;
-}
-
-.nav-panel__main-content::-webkit-scrollbar-track:vertical,
-.nav-panel__main-content::-webkit-scrollbar-track:horizontal {
-  background-color: transparent;
-}
-
-.nav-panel__main-content::-webkit-scrollbar-thumb:vertical,
-.nav-panel__main-content::-webkit-scrollbar-thumb:vertical:hover,
-.nav-panel__main-content::-webkit-scrollbar-thumb:vertical:active {
-  background-color: var(--scrollbar-color);
-  border: none;
-}
-
-.nav-panel__main-content::-webkit-scrollbar:vertical {
-  width: 6px !important
 }
 
 .qr-code {
@@ -2063,262 +1347,6 @@ html::-webkit-scrollbar,
   flex-shrink: 0;
 }
 
-/* Fix for jumping content */
-/* .custom-scrollbar {
-}
-.custom-scrollbar:hover {
-  padding-right: 10px !important;
-} */
-
-/* Simplebar scrollbars */
-/* .simplebar-scrollbar:before {
-  background: var(--scrollbar-color) !important;
-  border-radius: 1px !important;
-}
-
-.simplebar-track.simplebar-vertical.simplebar-hover .simplebar-scrollbar:before {
-  border-radius: 4px !important;
-}
-
-.simplebar-track.simplebar-vertical {
-  width: 5px !important;
-  margin-right: 4px;
-  transition: all 0.2s;
-  overflow: auto;
-}
-
-.simplebar-track.simplebar-vertical.simplebar-hover {
-  width: 12px !important;
-  margin-right: 2px;
-  transition: all 0.2s;
-} */
-
-/* THEMES | overlay-scrollbars */
-/* os-theme-minimal-light */
-.os-theme-minimal-light
-   > .os-scrollbar {
-    padding: 0px;
-  }
-
-.os-theme-minimal-light
-   > .os-scrollbar-horizontal {
-    right: 16px;
-    height: 16px;
-  }
-
-.os-theme-minimal-light
-   > .os-scrollbar-vertical {
-    bottom: 16px;
-    width: 16px;
-  }
-
-.os-theme-minimal-light.os-host-rtl
-   > .os-scrollbar-horizontal {
-    left: 16px;
-    right: 0;
-  }
-
-.os-theme-minimal-light
-   > .os-scrollbar-corner {
-    height: 16px;
-    width: 16px;
-    background-color: transparent;
-  }
-
-.os-theme-minimal-light
-   > .os-scrollbar
-   > .os-scrollbar-track,
-.os-theme-minimal-light
-   > .os-scrollbar
-   > .os-scrollbar-track
-   > .os-scrollbar-handle {
-    background: transparent;
-  }
-
-.os-theme-minimal-light
-   > .os-scrollbar
-   > .os-scrollbar-track
-   > .os-scrollbar-handle:before {
-    content: '';
-    display: block;
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    left: 0;
-    top: 0;
-    background: rgb(56, 58, 64);
-    opacity: 0.6;
-  }
-
-.os-theme-minimal-light
-   > .os-scrollbar
-   > .os-scrollbar-track
-   > .os-scrollbar-handle:hover:before {
-    opacity: 0.7;
-  }
-
-.os-theme-minimal-light
-   > .os-scrollbar
-   > .os-scrollbar-track
-   > .os-scrollbar-handle.active:before {
-    opacity: 0.8;
-  }
-
-.os-theme-minimal-light
-   > .os-scrollbar-horizontal
-   > .os-scrollbar-track
-   > .os-scrollbar-handle {
-    min-width: 32px;
-  }
-
-.os-theme-minimal-light
-   > .os-scrollbar-vertical
-   > .os-scrollbar-track
-   > .os-scrollbar-handle {
-    min-height: 32px;
-  }
-
-.os-theme-minimal-light
-   > .os-scrollbar-horizontal
-   > .os-scrollbar-track
-   > .os-scrollbar-handle:before {
-    height: 6px;
-    bottom: 0;
-    top: auto;
-  }
-
-.os-theme-minimal-light
-   > .os-scrollbar-vertical
-   > .os-scrollbar-track
-   > .os-scrollbar-handle:before {
-    margin-right: 4px;
-    width: 6px;
-    right: 0;
-    left: auto;
-  }
-
-.os-theme-minimal-light.os-host-rtl
-   > .os-scrollbar-vertical
-   > .os-scrollbar-track
-   > .os-scrollbar-handle:before {
-    left: 0;
-    right: auto;
-  }
-
-.os-theme-minimal-light
-   > .os-scrollbar-horizontal
-   > .os-scrollbar-track
-   > .os-scrollbar-handle:hover:before,
-.os-theme-minimal-light
-   > .os-scrollbar-horizontal
-   > .os-scrollbar-track
-   > .os-scrollbar-handle.active:before {
-    height: 100%;
-  }
-
-.os-theme-minimal-light
-   > .os-scrollbar-vertical
-   > .os-scrollbar-track
-   > .os-scrollbar-handle:hover:before,
-.os-theme-minimal-light
-   > .os-scrollbar-vertical
-   > .os-scrollbar-track
-   > .os-scrollbar-handle.active:before {
-    width: 100%;
-  }
-
-.os-theme-minimal-light.os-host-transition
-   > .os-scrollbar-horizontal
-   > .os-scrollbar-track
-   > .os-scrollbar-handle:before {
-    transition: opacity 0.3s, height 0.3s;
-  }
-
-.os-theme-minimal-light.os-host-transition
-   > .os-scrollbar-vertical
-   > .os-scrollbar-track
-   > .os-scrollbar-handle:before {
-    transition: opacity 0.3s, width 0.3s;
-  }
-
-/* Style overrides */
-i {
-  font-style: unset;
-}
-.v-list {
-  padding: 0px !important;
-}
-
-.v-card__text {
-  color: unset !important;
-}
-
-.v-toolbar__content {
-  padding: 0px !important;
-}
-
-/* .v-toolbar__content .v-btn.v-btn--icon.v-size--default { */
-.v-btn.v-btn--icon.v-size--default {
-  width: var(--window-toolbar-height) !important;
-  height: var(--window-toolbar-height) !important;
-}
-
-#window-toolbar
-  .window-toolbar__content--main,
-#action-toolbar
-  .v-toolbar__content {
-    padding: 0px 8px 0px 8px !important;
-  }
-
-/* .remove-padding-top {
-  padding-top: 0 !important;
-} */
-
-.v-btn.button-1 {
-  background: var(--button-bg-color-1) !important;
-  color: var(--button-color-1) !important;
-  box-shadow: none;
-}
-
-.v-btn.button-2 {
-  background: var(--button-bg-color-2) !important;
-  color: var(--button-color-2) !important;
-  box-shadow: none;
-}
-
-.text--title-1 {
-  font-size: 28px;
-  color: var(--title-color-1);
-}
-
-.text--title-2 {
-  font-size: 20px;
-  color: var(--title-color-1);
-}
-
-.text--sub-title-1 {
-  font-size: 14px;
-  margin-bottom: 8px;
-  color: var(--sub-title-color-1);
-  text-transform: uppercase;
-  user-select: none;
-}
-
-.text--sub-title-2 {
-  font-size: 14px;
-  margin-bottom: 8px;
-  color: var(--color-5);
-}
-
-.content__description {
-  font-size: 16px;
-  margin-bottom: 8px;
-  line-height: 1.5;
-  word-break: break-word;
-  color: var(--sub-title-color-1);
-}
-
-/* ==== GENERAL COMPONENTS ==== */
 .overlay--drag-over {
   position: absolute;
   width: 100%;
@@ -2335,11 +1363,6 @@ i {
   transition: all 0.1s;
 }
 
-.app-content {
-  background-color: var(--app-content-bg-color);
-  padding-top: var(--header-height) !important;
-}
-
 .list-item__checkbox {
   z-index: 3;
   right: 0;
@@ -2351,7 +1374,6 @@ i {
   opacity: 1;
 }
 
-/* Dir-item */
 .dir-item__checkbox {
   z-index: 3;
   opacity: 0;
@@ -2364,22 +1386,6 @@ i {
   opacity: 1;
 }
 
-.custom-divider {
-  padding: 18px 0px;
-  border-top: 1px solid;
-  width: 100%;
-  border-width: thin;
-  border-color: var(--highlight-color-2)
-}
-
-.custom-divider[vertical] {
-  margin: 0px 4px;
-  height: 20px;
-  width: 2px;
-  background-color: var(--highlight-color-2)
-}
-
-/* Content area */
 .content-area {
   padding: 8px 24px;
   height: calc(100vh - var(--window-toolbar-height) - var(--action-toolbar-height));
@@ -2392,7 +1398,6 @@ i {
   color: var(--title-color-1);
 }
 
-/* Content card */
 .content-card {
   padding: 20px 24px;
   margin: 16px 0px;
@@ -2400,153 +1405,9 @@ i {
   box-shadow: 0px 8px 32px rgb(0, 0, 0, 0.1);
 }
 
-.theme--light.v-btn.v-btn--disabled {
-  opacity: 0.5;
-}
-
-.theme--light.v-btn.v-btn--disabled
-  .v-icon.css-override {
-    color: var(--icon-color-2) !important;
-}
-
-/* inputs */
-.v-input .v-messages {
-  font-size: 14px;
-}
-
-.v-input:not(.v-input--is-disabled)
-  input,
-.v-input:not(.v-input--is-disabled)
-  textarea {
-    color: var(--color-5) !important;
-  }
-
-/* .v-input:not(.v-input--is-disabled) input,
-.v-input:not(.v-input--is-disabled) textarea {
-  color: var(--color-6) !important;
-} */
-
-.v-input
-  .v-label {
-    color: var(--radio-text-color-1) !important;
-}
-
-.v-input--is-label-active
-  .v-input--switch__track {
-    background: var(--input_active-track-color-1) !important;
-}
-
-.v-input--is-label-active
-  .v-input--switch__thumb {
-    background: var(--input_active-thumb-color-1) !important;
-  }
-
-.v-input--switch__track {
-  background: var(--input-track-color-1) !important;
-}
-
-.v-input--switch__thumb {
-  background: var(--input-thumb-color-1) !important;
-}
-
-.v-text-field__slot
-  label.v-label--active {
-    color: var(--color-1) !important;
-  }
-
-.v-text-field__slot {
-  caret-color: var(--color-1) !important
-}
-
-.theme--light.v-text-field >
-  .v-input__control >
-    .v-input__slot:before {
-        border-color: var(--color-8) !important;
-    }
-
-.v-input.v-input--is-focused
-  * {
-    color: var(--color-5) !important;
-  }
-
-.v-input.v-input--is-focused
-  .v-messages__message {
-    color: var(--color-5) !important;
-  }
-
-.v-input.error--text
-  .v-messages__message {
-    color: #ef5350 !important;
-  }
-
-.v-input.v-input--is-disabled
-  input {
-    color: var(--color-7) !important;
-  }
-
-.v-chip--select.v-chip
-  * {
-    color: var(--color-5) !important;
-  }
-
-.v-chip.v-chip--active {
-    background: var(--highlight-color-2) !important;
-}
-
-.v-chip:not(.v-chip--active) {
-    background: var(--highlight-color-3) !important;
-}
-
-/* v-text-field */
-.v-input.v-input--is-label-active.v-text-field.v-select
-  * {
-    color: var(--color-6) !important;
-  }
-
-.v-input.v-text-field
-  .v-label {
-    color: var(--color-7) !important;
-  }
-
-.v-input.v-text-field.v-input--is-focused
-  .v-label {
-    color: var(--color-5) !important;
-  }
-
 .workspace-action__card {
   padding: 12px 0px;
   margin: 12px 0px;
-}
-
-.v-input.theme--light.v-input--selection-controls
-  .v-input--selection-controls__input
-    * {
-      color: var(--key-color-1) !important;
-    }
-
-.v-input.theme--light.v-input--selection-controls
-  .v-input--selection-controls__input
-    i.mdi-radiobox-blank {
-      color: var(--color-7) !important;
-    }
-
-.v-application--is-ltr
-  .v-input--selection-controls__input {
-    margin-right: 12px !important;
-  }
-
-.v-input__slider .v-slider__thumb,
-.v-input__slider .v-slider__thumb-container-fill {
-  color: var(--key-color-1) !important;
-  background: var(--key-color-1) !important;
-  background-color: var(--key-color-1) !important
-}
-
-.v-input__slider .v-slider__track-background,
-.v-input__slider .v-slider__track-fill {
-  color: var(--highlight-color-2) !important;
-  background: var(--highlight-color-2) !important;
-  background-color: var(--highlight-color-2) !important
 }
 
 .button-container {
@@ -2572,20 +1433,11 @@ i {
 
 .tab-view
   .tab-view__header {
-    position: sticky;
-    top: 0px;
-    height: fit-content;
     z-index: 2;
-    background-color: var(--bg-color-1);
-  }
-
-.tab-view
-  .tab-view__header {
     position: sticky;
     height: fit-content;
     top: -8px;
     background-color: var(--bg-color-1);
-    z-index: 2;
   }
   @media (min-width: 700px) {
     .tab-view
@@ -2655,258 +1507,5 @@ i {
 .v-tabs .v-slide-group__next .v-icon,
 .v-tabs .v-slide-group__prev .v-icon {
   color: rgb(255, 255, 255, 0.3) !important
-}
-
-/* Transitions */
-@keyframes fade-in-slide-up {
-  from { opacity: 0; transform: translateY(10px); }
-  to { opacity: 1; }
-}
-
-@keyframes fade-in {
-  from { opacity: 0; }
-  to { opacity: 1; }
-}
-
-@keyframes scale-in {
-  from { transform: scale(0.95) }
-  to { transform: scale(1) }
-}
-
-@keyframes scale-120-percent-to-initial {
-  from { transform: scale(1.2) }
-  to { transform: scale(1) }
-}
-
-.fade-in-slide-up-500ms {
-  animation: fade-in-slide-up 500ms;
-}
-
-.fade-in-200ms {
-  animation: fade-in 200ms;
-}
-
-.fade-in-500ms {
-  animation: fade-in 500ms;
-}
-
-.fade-in-1s {
-  animation: fade-in 1s;
-}
-
-.fade-in-2s {
-  animation: fade-in 2s;
-}
-
-.fade-scale-in-500ms {
-  animation: fade-in 500ms, scale-in 500ms;
-}
-
-.fade-in-2s-scale-in-500ms {
-  animation: fade-in 2s, scale-in 500ms;
-}
-
-.fade-in-2000ms-scale-120-percent-to-initial-2000ms {
-  animation: fade-in 2000ms, scale-120-percent-to-initial 2000ms;
-}
-
-.route-transition-enter-active {
-  transition: all 0.3s ease;
-}
-
-.route-transition-leave-active {
-  transition: all 0.3s ease;
-}
-
-.route-transition-enter,
-.route-transition-leave-to {
-  opacity: 0;
-}
-
-.context-menu-transition-enter-active {
-  transition: all 0.2s ease;
-}
-
-.context-menu-transition-leave-active {
-  transition: all 0s;
-}
-
-.context-menu-transition-enter,
-.context-menu-transition-leave-to {
-  opacity: 0;
-  transform:
-    scale(0.1)
-    /* Prevent tooltip showing up during animation*/
-    translateX(16px);
-}
-
-.context-sub-menu-transition-enter-active {
-  transition: all 0.2s ease;
-}
-
-.context-sub-menu-transition-leave-active {
-  transition: all 0.2s ease;
-}
-
-.context-sub-menu-transition-enter {
-  transform: translateX(20px);
-  opacity: 0;
-}
-
-.context-sub-menu-transition-leave-to {
-  transform: translateX(-20px);
-  opacity: 0;
-}
-
-.context-sub-menu-transition-reversed-enter-active {
-  transition: all 0.2s ease;
-}
-
-.context-sub-menu-transition-reversed-leave-active {
-  transition: all 0.2s ease;
-}
-
-.context-sub-menu-transition-reversed-enter {
-  transform: translateX(-20px);
-  opacity: 0;
-}
-
-.context-sub-menu-transition-reversed-leave-to {
-  transform: translateX(20px);
-  opacity: 0;
-}
-
-.fade-in-enter-active {
-  transition: all 0.2s ease;
-}
-
-.fade-in-leave-active {
-  transition: all 0.2s ease;
-}
-
-.fade-in-enter,
-.fade-in-leave-to {
-  transition: all 0.2s ease;
-  opacity: 0;
-}
-
-.fade-in-1s-enter-active {
-  transition: all 1s ease;
-}
-
-.fade-in-1s-leave-active {
-  transition: all 1s ease;
-}
-
-.fade-in-1s-enter,
-.fade-in-1s-leave-to {
-  opacity: 0;
-}
-
-.slide-fade-left-enter-active {
-  transition: all 0.2s ease;
-}
-
-.slide-fade-left-leave-active {
-  transition: all 0.2s ease;
-}
-
-.slide-fade-left-enter,
-.slide-fade-left-leave-to {
-  transform: translateX(-10px);
-  opacity: 0;
-}
-
-.group-fade-in-enter-active,
-.group-fade-in-leave-active {
-  transition: all 0.5s;
-}
-
-.group-fade-in-enter,
-.group-fade-in-leave-to {
-  position: absolute;
-  opacity: 0;
-}
-
-.group-slide-fade-left-enter-active,
-.group-slide-fade-left-leave-active {
-  transition: all 0.5s;
-}
-
-.group-slide-fade-left-enter,
-.group-slide-fade-left-leave-to {
-  position: absolute;
-  transform: translateX(-10px);
-  opacity: 0;
-}
-
-.slide-fade-up-enter-active,
-.slide-fade-up-leave-active {
-  transition: all 1.5s;
-}
-
-.slide-fade-up-enter,
-.slide-fade-up-leave-to {
-  transition: all 1.5s;
-  transform: translateY(16px);
-  opacity: 0;
-}
-
-.slide-fade-up-300ms-enter-active,
-.slide-fade-up-300ms-leave-active {
-  transition: all 0.3s;
-}
-
-.slide-fade-up-300ms-enter,
-.slide-fade-up-300ms-leave-to {
-  transition: all 0.3s;
-  transform: translateY(16px);
-  opacity: 0;
-}
-
-.slide-fade-down-500ms-enter-active,
-.slide-fade-down-500ms-leave-active {
-  transition: all 0.5s;
-}
-
-.slide-fade-down-500ms-enter,
-.slide-fade-down-500ms-leave-to {
-  transition: all 0.5s;
-  transform: translateY(-16px);
-  opacity: 0;
-}
-
-.slide-fade-down-300ms-enter-active,
-.slide-fade-down-300ms-leave-active {
-  transition: all 0.3s;
-}
-
-.slide-fade-down-300ms-enter,
-.slide-fade-down-300ms-leave-to {
-  transition: all 0.3s;
-  transform: translateY(-36px);
-  opacity: 0;
-}
-
-.overlay--item-drag__transition-enter-active {
-  /* Remove enter animation to
-  prevent position change lag */
-  transition: all 0s;
-}
-
-.overlay--item-drag__transition-leave-active {
-  transition: all 0.3s;
-}
-
-.overlay--item-drag__transition-enter {
-  transition: all 0.3s ease;
-  transform: scale(0.6);
-  opacity: 0;
-}
-
-.overlay--item-drag__transition-leave-to {
-  transition: all 0.3s ease;
-  transform: scale(0.6) translateX(-64px) translateY(-32px);
-  opacity: 0;
 }
 </style>
