@@ -13,9 +13,10 @@ Copyright © 2021 - present Aleksey Hoffman. All rights reserved.
       :position-y="contextMenus.dirItem.y"
       :close-on-content-click="false"
       offset-y
+      offset-overflow
       absolute
     >
-      <div class="context-menu__container">
+      <div class="context-menu__container fade-in-1s">
         <transition
           :name="contextMenus.dirItem.subMenu.value
             ? 'context-sub-menu-transition'
@@ -23,13 +24,16 @@ Copyright © 2021 - present Aleksey Hoffman. All rights reserved.
           mode="out-in"
         >
           <div key="mainMenu" v-if="!contextMenus.dirItem.subMenu.value">
-            <v-list dense>
+            <v-list
+              v-if="toolbarItems.length > 0"
+              dense
+            >
               <!-- context-menu::main-view::toolbar -->
               <div class="context-menu__toolbar">
                 <v-tooltip
                   v-for="(item, index) in filteredList(toolbarItems)"
-                  :key="'toolbar-item-' + index"
                   open-delay="200"
+                  :key="'toolbar-item-' + index"
                   max-width="300px"
                   bottom
                 >
@@ -46,33 +50,108 @@ Copyright © 2021 - present Aleksey Hoffman. All rights reserved.
                       </v-icon>
                     </v-btn>
                   </template>
-                  <span>
+                  <span v-if="item.tooltip && item.tooltip.shortcutList">
+                    <div
+                      class="tooltip__shortcut-list-item"
+                      v-for="(shortcutItem, index) in item.tooltip.shortcutList"
+                      :key="`shortcut-list-item-${index}`"
+                    >
+                      <div class="tooltip__shortcut-list-item__title tooltip__description">
+                        {{shortcutItem.title}}
+                      </div>
+                      <span
+                        class="tooltip__shortcut-list__shortcut inline-code--light"
+                        v-html="shortcutItem.shortcut"
+                      ></span>
+                    </div>
+                  </span>
+
+                  <span v-if="item.tooltip && item.tooltip.modifierList">
+                    <div
+                      class="tooltip__modifier-list-item"
+                      v-for="(modifierItem, index) in item.tooltip.modifierList"
+                      :key="`modifier-list-item-${index}`"
+                    >
+                      <div class="tooltip__modifier-list-item__title tooltip__description">
+                        <span
+                          class="tooltip__modifier-list__modifier inline-code--light"
+                          v-if="modifierItem.modifier"
+                          v-html="modifierItem.modifier"
+                        ></span>
+                        {{modifierItem.title}}
+                      </div>
+                      <div
+                        class="tooltip__modifier-list__title tooltip__description"
+                        v-if="!modifierItem.modifier"
+                      >
+                        Modifiers:
+                      </div>
+                    </div>
+                  </span>
+
+                  <span v-if="item.tooltip && !item.tooltip.shortcutList && !item.tooltip.modifierList">
                     <div class="tooltip__description">{{item.tooltip.title}}</div>
                     <div class="tooltip__shortcut" v-html="item.tooltip.text"></div>
                   </span>
                 </v-tooltip>
               </div>
             </v-list>
-            <v-divider></v-divider>
+
+            <v-divider
+              v-if="toolbarItems.length > 0"
+            ></v-divider>
 
             <!-- context-menu::main-view::list -->
             <v-list dense>
-              <v-list-item
+              <v-tooltip
                 v-for="(item, index) in filteredList(menuItems)"
                 :key="'menu-item-' + index"
-                @click.stop="runOnClickEvent(item)"
+                :disabled="!item.tooltip"
+                open-delay="200"
+                bottom
               >
-                <v-list-item-icon>
-                  <v-icon :size="item.iconSize">
-                    {{item.icon}}
-                  </v-icon>
-                </v-list-item-icon>
-                <v-list-item-content>
-                  <v-list-item-title>
-                    {{item.title}}
-                  </v-list-item-title>
-                </v-list-item-content>
-              </v-list-item>
+                <template v-slot:activator="{on}">
+                  <v-list-item
+                    v-on="on"
+                    @click.stop="runOnClickEvent(item)"
+                  >
+                    <v-list-item-icon>
+                      <v-icon :size="item.iconSize">
+                        {{item.icon}}
+                      </v-icon>
+                    </v-list-item-icon>
+                    <v-list-item-content>
+                      <v-list-item-title>
+                        {{item.title}}
+                      </v-list-item-title>
+                    </v-list-item-content>
+                    <v-list-item-icon>
+                      <v-icon
+                        v-if="item.subMenu"
+                        size="24px"
+                        is-sub-menu
+                      >
+                        mdi-chevron-right
+                      </v-icon>
+                    </v-list-item-icon>
+                  </v-list-item>
+                </template>
+                <span v-if="item.tooltip && item.tooltip.shortcutList">
+                  <div
+                    class="tooltip__shortcut-list-item"
+                    v-for="(shortcutItem, index) in item.tooltip.shortcutList"
+                    :key="`shortcut-list-item-${index}`"
+                  >
+                    <div class="tooltip__shortcut-list-item__title tooltip__description">
+                      {{shortcutItem.title}}
+                    </div>
+                    <span
+                      class="tooltip__shortcut-list__shortcut inline-code--light"
+                      v-html="shortcutItem.shortcut"
+                    ></span>
+                  </div>
+                </span>
+              </v-tooltip>
             </v-list>
           </div>
 
@@ -201,28 +280,20 @@ Copyright © 2021 - present Aleksey Hoffman. All rights reserved.
 </template>
 
 <script>
-import { mapFields } from 'vuex-map-fields'
-import { shell } from 'electron'
+import {mapFields} from 'vuex-map-fields'
+import {shell} from 'electron'
 const childProcess = require('child_process')
 const PATH = require('path')
 
 export default {
-  data () {
-    return {
-    }
-  },
   mounted () {
     // Init listeners
     this.$eventHub.$on('openInTerminal', () => {
-      this.$store.dispatch('OPEN_SELECTED_IN_TERMINAL', { asAdmin: false })
+      this.$store.dispatch('OPEN_SELECTED_IN_TERMINAL', {asAdmin: false})
     })
 
     this.$eventHub.$on('openInTerminalAsAdmin', () => {
-      this.$store.dispatch('OPEN_SELECTED_IN_TERMINAL', { asAdmin: true })
-    })
-
-    this.$eventHub.$on('openWithQuickView', () => {
-      this.$store.dispatch('OPEN_WITH_QUICK_VIEW')
+      this.$store.dispatch('OPEN_SELECTED_IN_TERMINAL', {asAdmin: true})
     })
 
     this.$eventHub.$on('openInNativeFileManager', () => {
@@ -241,7 +312,7 @@ export default {
       if (!value) {
         this.contextMenus.dirItem.subMenu.value = false
       }
-    }
+    },
   },
   computed: {
     ...mapFields({
@@ -252,7 +323,7 @@ export default {
       pinnedItems: 'storageData.pinned.items',
       protectedItems: 'storageData.protected.items',
       externalProgramsDefaultItems: 'storageData.settings.externalPrograms.defaultItems',
-      externalProgramsCustomItems: 'storageData.settings.externalPrograms.items'
+      externalProgramsCustomItems: 'storageData.settings.externalPrograms.items',
     }),
     targetItems () {
       return this.$store.state.contextMenus.dirItem.targetItems
@@ -260,7 +331,7 @@ export default {
     targetItemsStats () {
       return this.$store.state.contextMenus.dirItem.targetItemsStats
     },
-    toolbarItems () {
+    dirItemToolbarItems () {
       return [
         {
           title: 'protect',
@@ -278,8 +349,8 @@ export default {
               If enabled, the item will be added to the protected list. 
               It will be protected from being modified / deleted from within this app only.
               Other apps can still modify / delete it.
-            `
-          }
+            `,
+          },
         },
         {
           title: 'pin',
@@ -291,7 +362,10 @@ export default {
           },
           icon: 'mdi-pin-outline',
           iconSize: '20px',
-          tooltip: { title: 'Add to pinned', text: '' }
+          tooltip: {
+            title: 'Add to pinned',
+            text: 'Pinned items can be found on the dashboard page',
+          },
         },
         {
           title: 'rename',
@@ -305,9 +379,13 @@ export default {
           icon: 'mdi-form-textbox',
           iconSize: '20px',
           tooltip: {
-            title: 'Rename selected',
-            text: this.shortcuts.renameSelected.shortcut
-          }
+            shortcutList: [
+              {
+                title: 'Rename selected',
+                shortcut: this.shortcuts.renameSelected.shortcut,
+              },
+            ],
+          },
         },
         {
           title: 'move',
@@ -315,15 +393,23 @@ export default {
           targetTypes: ['directory', 'file', 'file-symlink', 'directory-symlink'],
           isActive: false,
           onClick: () => {
-            this.$store.dispatch('SET_FS_CLIPBOARD', { type: 'move' })
+            this.$store.dispatch('SET_TO_FS_CLIPBOARD', {type: 'move'})
           },
           closesMenu: true,
           icon: 'mdi-content-duplicate',
           iconSize: '18px',
           tooltip: {
-            title: 'Move selected to another directory',
-            text: this.shortcuts.moveSelectedDirItems.shortcut
-          }
+            shortcutList: [
+              {
+                title: 'Set selected items for moving',
+                shortcut: this.shortcuts.setDirItemsForMoving.shortcut,
+              },
+              {
+                title: 'Add selected items for moving',
+                shortcut: this.shortcuts.addDirItemsForMoving.shortcut,
+              },
+            ],
+          },
         },
         {
           title: 'copy',
@@ -331,15 +417,23 @@ export default {
           targetTypes: ['directory', 'file', 'file-symlink', 'directory-symlink'],
           isActive: false,
           onClick: () => {
-            this.$store.dispatch('SET_FS_CLIPBOARD', { type: 'copy' })
+            this.$store.dispatch('SET_TO_FS_CLIPBOARD', {type: 'copy'})
           },
           closesMenu: true,
           icon: 'far fa-copy',
           iconSize: '18px',
           tooltip: {
-            title: 'Copy selected',
-            text: this.shortcuts.copySelectedDirItems.shortcut
-          }
+            shortcutList: [
+              {
+                title: 'Set selected items for copying',
+                shortcut: this.shortcuts.setDirItemsForCopying.shortcut,
+              },
+              {
+                title: 'Add selected items for copying',
+                shortcut: this.shortcuts.addDirItemsForCopying.shortcut,
+              },
+            ],
+          },
         },
         {
           title: 'link',
@@ -348,12 +442,42 @@ export default {
           disallowedExtensions: ['lnk'],
           isActive: false,
           onClick: () => {
-            this.$store.dispatch('MAKE_DIR_ITEM_LINK')
+            let srcPath = this.targetItems[0].path
+            let params = {
+              title: '',
+              destPath: srcPath,
+            }
+            let linkModifier = this.linkModifier
+            if (linkModifier === 'windows-link') {
+              params.title = 'Create Windows link (.lnk)'
+              params.destPath = `${srcPath}.lnk`
+            }
+            else {
+              params.title = 'Create symlink'
+            }
+            this.$store.dispatch('SHOW_CONFIRMATION_DIALOG_MAKE_LINK', params)
+              .then(result => {
+                let destPath = result.data.inputs[0].model
+                if (linkModifier === 'windows-link') {
+                  this.$store.dispatch('MAKE_DIR_ITEM_LINK', {
+                    linkType: 'windows-link',
+                    srcPath,
+                    destPath,
+                  })
+                }
+                else {
+                  this.$store.dispatch('MAKE_DIR_ITEM_LINK', {
+                    linkType: 'symlink',
+                    srcPath,
+                    destPath,
+                  })
+                }
+              })
           },
           closesMenu: true,
-          icon: 'mdi-link-variant-plus',
+          icon: this.linkActionIcon,
           iconSize: '22px',
-          tooltip: { title: 'Create symbolic link (shortcut)', text: '' }
+          tooltip: this.linkActionTooltip,
         },
         {
           title: 'trash',
@@ -362,31 +486,49 @@ export default {
           isActive: false,
           onClick: () => {
             if (this.inputState.shift) {
-              this.$store.dispatch('INIT_DELETE_SELECTED')
+              this.$store.dispatch('DELETE_SELECTED_DIR_ITEMS')
             }
             else {
-              this.$store.dispatch('TRASH_SELECTED')
+              this.$store.dispatch('TRASH_SELECTED_DIR_ITEMS')
             }
           },
           closesMenu: true,
-          icon: 'mdi-trash-can-outline',
+          icon: this.inputState.shift
+            ? 'mdi-delete-forever-outline'
+            : 'mdi-trash-can-outline',
           iconSize: '22px',
           tooltip: {
-            title: 'Move selected to trash',
-            text: `
-              ${this.shortcuts.trashSelected.description}
-              <br>${this.shortcuts.trashSelected.shortcut}
-              <br>${this.shortcuts.deleteSelected.description}
-              <br>${this.shortcuts.deleteSelected.shortcut}
-            `
-          }
-        }
+            shortcutList: [
+              {
+                title: this.shortcuts.trashSelected.description,
+                shortcut: this.shortcuts.trashSelected.shortcut,
+              },
+              {
+                title: this.shortcuts.deleteSelected.description,
+                shortcut: this.shortcuts.deleteSelected.shortcut,
+              },
+            ],
+          },
+        },
       ]
     },
-    menuItems () {
+    toolbarItems () {
+      if (this.contextMenus.dirItem.targetType === 'userDir') {
+        return []
+      }
+      else if (this.contextMenus.dirItem.targetType === 'drive') {
+        return []
+      }
+      else {
+        return this.dirItemToolbarItems
+      }
+    },
+    dirItemMenuItems () {
       return [
         {
+          name: 'open-with',
           title: 'Open with',
+          subMenu: 'open-with',
           selectionType: ['single', 'multiple'],
           targetTypes: ['directory', 'file', 'file-symlink', 'directory-symlink'],
           onClick: () => {
@@ -397,9 +539,31 @@ export default {
           },
           closesMenu: false,
           icon: 'mdi-subdirectory-arrow-right',
-          iconSize: '20px'
+          iconSize: '20px',
         },
         {
+          name: 'quick-view',
+          title: 'Quick view',
+          selectionType: ['single'],
+          targetTypes: ['file', 'file-symlink'],
+          onClick: () => {
+            this.$store.dispatch('OPEN_WITH_QUICK_VIEW')
+          },
+          closesMenu: true,
+          icon: 'mdi-text-box-search-outline',
+          iconSize: '18px',
+          tooltip: {
+            shortcutList: [
+              {
+                name: 'review file in a separate window',
+                title: 'Preview file in a separate window',
+                shortcut: this.shortcuts.openWithQuickView.shortcut,
+              },
+            ],
+          },
+        },
+        {
+          name: 'new-tab',
           title: 'New tab',
           selectionType: ['single'],
           targetTypes: ['directory', 'directory-symlink'],
@@ -408,35 +572,53 @@ export default {
           },
           closesMenu: true,
           icon: 'mdi-tab',
-          iconSize: '18px'
+          iconSize: '18px',
         },
         {
+          name: 'edit-tags',
+          title: 'Edit tags',
+          selectionType: ['single', 'multiple'],
+          targetTypes: ['file', 'directory', 'file-symlink', 'directory-symlink'],
+          onClick: () => {
+            this.contextMenus.dirItem.subMenu.target = 'edit-tags'
+            this.contextMenus.dirItem.subMenu.title = 'Edit tags'
+            this.contextMenus.dirItem.subMenu.value = true
+            this.adjustMenuPositionToSubMenu()
+          },
+          closesMenu: false,
+          icon: 'mdi-tag-text-outline',
+          iconSize: '18px',
+        },
+        {
+          name: 'share-directory-with-local-devices',
           title: 'Share directory with local devices',
           targetTypes: ['directory', 'directory-symlink'],
           selectionType: ['single'],
           onClick: () => {
-            this.$eventHub.$emit('app:method', {
-              method: 'initLocalDirectoryShare'
+            this.$eventHub.$emit('fsLocalServerManager:method', {
+              method: 'initLocalDirectoryShare',
             })
           },
           closesMenu: true,
-          icon: 'mdi-folder-search-outline',
-          iconSize: '18px'
+          icon: 'mdi-folder-network-outline',
+          iconSize: '18px',
         },
         {
+          name: 'share-file-with-local-devices',
           title: 'Share file with local devices',
           targetTypes: ['file'],
           selectionType: ['single'],
           onClick: () => {
-            this.$eventHub.$emit('app:method', {
-              method: 'initLocalFileShare'
+            this.$eventHub.$emit('fsLocalServerManager:method', {
+              method: 'initLocalFileShare',
             })
           },
           closesMenu: true,
-          icon: 'mdi-share-variant',
-          iconSize: '18px'
+          icon: 'mdi-access-point-network',
+          iconSize: '18px',
         },
         {
+          name: 'compress-to-archive',
           title: 'Compress to archive',
           targetTypes: ['file', 'directory', 'file-symlink', 'directory-symlink'],
           selectionType: ['single', 'multiple'],
@@ -446,9 +628,10 @@ export default {
           },
           closesMenu: true,
           icon: 'mdi-cube-outline',
-          iconSize: '20px'
+          iconSize: '20px',
         },
         {
+          name: 'extract-archive',
           title: 'Extract archive',
           targetTypes: ['file'],
           selectionType: ['single'],
@@ -458,22 +641,24 @@ export default {
           },
           closesMenu: true,
           icon: 'mdi-package-variant',
-          iconSize: '20px'
+          iconSize: '20px',
         },
         {
+          name: 'permissions',
           title: 'Permissions',
           targetTypes: ['file', 'directory'],
           selectionType: ['single'],
           onClick: () => {
-            this.$store.dispatch('INIT_FETCH_CONTEXT_MENU_TARGET_ITEMS', { type: 'dirItem' })
+            this.$store.dispatch('INIT_FETCH_CONTEXT_MENU_TARGET_ITEMS', {type: 'dirItem'})
             this.dialogs.dirItemPermissionManagerDialog.value = true
           },
           closesMenu: true,
           icon: 'mdi-lock-outline',
-          iconSize: '18px'
-        }
+          iconSize: '18px',
+        },
         // TODO: finish in v1.2.0
         // {
+        //   name: 'convert-or-edit-image',
         //   title: 'Convert or edit image',
         //   targetTypes: ['file'],
         //   allowedFilesTypes: ['image'],
@@ -487,6 +672,7 @@ export default {
         //   iconSize: '23px'
         // },
         // {
+        //   name: 'convert-or-edit-audio',
         //   title: 'Convert or edit audio',
         //   targetTypes: ['file'],
         //   allowedFilesTypes: ['audio'],
@@ -500,6 +686,7 @@ export default {
         //   iconSize: '23px'
         // },
         // {
+        //   name: 'convert-or-edit-video',
         //   title: 'Convert or edit video',
         //   targetTypes: ['file'],
         //   allowedFilesTypes: ['video'],
@@ -513,6 +700,7 @@ export default {
         //   iconSize: '23px'
         // },
         // {
+        //   name: 'calculate-hash',
         //   title: 'Calculate hash',
         //   selectionType: ['single'],
         //   targetTypes: ['file', 'file-symlink'],
@@ -525,6 +713,39 @@ export default {
         //   iconSize: '22px'
         // }
       ]
+    },
+    menuItems () {
+      if (this.contextMenus.dirItem.targetType === 'userDir') {
+        return [
+          ...this.dirItemMenuItems,
+          ...[
+            {
+              title: 'Edit card',
+              selectionType: ['single'],
+              targetTypes: ['directory', 'file', 'file-symlink', 'directory-symlink'],
+              onClick: () => {
+                this.dialogs.userDirectoryEditorDialog.data = this.contextMenus.dirItem.targetData
+                this.dialogs.userDirectoryEditorDialog.value = true
+              },
+              closesMenu: false,
+              icon: 'mdi-pencil-outline',
+              iconSize: '20px',
+            },
+          ],
+        ]
+      }
+      else if (this.contextMenus.dirItem.targetType === 'drive') {
+        return [
+          ...this.dirItemMenuItemsDriveFiltered,
+        ]
+      }
+      else {
+        return this.dirItemMenuItems
+      }
+    },
+    dirItemMenuItemsDriveFiltered () {
+      return this.dirItemMenuItems
+        .filter(item => item.name !== 'permissions')
     },
     externalProgramsDefaultItemsFiltered () {
       const itemsFilteredByItemType = this.filterAllowedTargetTypes(this.externalProgramsDefaultItems)
@@ -556,7 +777,43 @@ export default {
         return everyDirItemMatchDisallowedExtensions
       })
       return itemsFilteredByDisallowed
-    }
+    },
+    linkModifier () {
+      if (this.inputState.alt) {
+        return 'windows-link'
+      }
+      else {
+        return 'symlink'
+      }
+    },
+    linkActionIcon () {
+      if (process.platform === 'win32') {
+        return this.linkModifier === 'windows-link'
+          ? 'mdi-link-box-variant-outline'
+          : 'mdi-link-variant-plus'
+      }
+      else {
+        return this.linkModifier === 'hard-link'
+          ? 'mdi-link-plus'
+          : 'mdi-link-variant-plus'
+      }
+    },
+    linkActionTooltip () {
+      let tooltip = {
+        modifierList: [
+          {
+            title: 'Create symbolic link',
+          },
+        ],
+      }
+      if (process.platform === 'win32') {
+        tooltip.modifierList.push({
+          title: 'Windows link (.lnk)',
+          modifier: 'Alt',
+        })
+      }
+      return tooltip
+    },
   },
   methods: {
     filteredList (list) {
@@ -672,7 +929,7 @@ export default {
         }
         this.$eventHub.$emit('app:method', {
           method: method,
-          params: item.onClick.params
+          params: item.onClick.params,
         })
       }
       if (item.closesMenu) {
@@ -705,14 +962,14 @@ export default {
               timeout: 8000,
               closeButton: true,
               title: 'Error: couldn\'t open the program:',
-              message: error
+              message: error,
             })
           }
-        }
+        },
       )
       this.contextMenus.dirItem.value = false
-    }
-  }
+    },
+  },
 }
 </script>
 
