@@ -6,9 +6,9 @@ Copyright © 2021 - present Aleksey Hoffman. All rights reserved.
 <template>
   <div>
     <v-navigation-drawer
-      class="info-panel"
       v-if="$vuetify.breakpoint.mdAndUp"
-      v-model="navigatorViewInfoPanel"
+      v-model="navigatorViewInfoPanel.value"
+      class="info-panel"
       app
       clipped
       floating
@@ -17,78 +17,65 @@ Copyright © 2021 - present Aleksey Hoffman. All rights reserved.
       touchless
       width="280"
     >
-      <v-layout
-        class="info-panel__preview-container"
-        align-center
-        column
-        justify-center
-      >
-        <v-icon
-          :class="previewIcon.class"
-          @click="() => {previewIcon.onClick()}"
-          :size="previewIcon.size"
-          :aria-label="previewIcon.ariaLabel"
-          :data-icon="previewIcon.dataIcon"
-        >{{previewIcon.name}}
-        </v-icon>
-
-        <!-- preview::media-elements -->
-        <div class="info-panel__preview-container__media-container">
-          <img
-            class="fade-in-1s info-panel__preview-container__media"
-            v-if="showPreviewItem('image') && !selectedDirItemData.itemProps.isInaccessible"
-            :src="previewMediaSrc"
+      <div class="info-panel__preview-container">
+        <div class="info-panel__preview-container__info-overlay">
+          <v-icon
+            class="info-panel__preview-container__info-overlay__item-type-icon"
+            :class="previewIcon.class"
+            :size="previewIcon.size"
+            :aria-label="previewIcon.ariaLabel"
+            :data-icon="previewIcon.dataIcon"
+            @click="previewIcon.onClick()"
           >
-          <audio
-            class="fade-in-1s info-panel__preview-container__media"
-            v-if="showPreviewItem('audio') && !selectedDirItemData.itemProps.isInaccessible"
-            :src="previewMediaSrc"
-            controls
-            preload="metadata"
-            style="display: none"
-          />
-          <video
-            class="fade-in-1s info-panel__preview-container__media"
-            v-if="showPreviewItem('video') && !selectedDirItemData.itemProps.isInaccessible"
-            :src="previewMediaSrc"
-            controls
-            preload="metadata"
-          />
-        </div>
+            {{previewIcon.name}}
+          </v-icon>
 
-        <!-- preview::mime -->
-        <div v-show="['file', 'file-symlink'].includes(infoPanelThumbType)">
-          {{$utils.getExt(selectedDirItemData.title)}}
-        </div>
-      </v-layout>
+          <div
+            v-show="['file', 'file-symlink'].includes(itemType)"
+            class="info-panel__preview-container__info-overlay__item-ext"
+            v-html="itemExt"
+          />
 
-      <!-- info-container -->
+          <div
+            v-show="itemMimeDescription === 'audio'"
+            class="info-panel__preview-container__info-overlay__audio-duration"
+          >
+            {{formattedInfoPanelAudioDuration}}
+          </div>
+        </div>
+        <div class="info-panel__preview-container__media-container" />
+      </div>
+
       <div class="info-panel__info-container">
-        <v-layout class="info-panel__header" column>
+        <v-layout
+          class="info-panel__header"
+          column
+        >
           <!-- info-container::header::title -->
           <v-tooltip bottom>
-            <template v-slot:activator="{ on }">
+            <template #activator="{on}">
               <div
-                v-on="on"
-                @click.ctrl="$store.dispatch(
-                  'COPY_TEXT_TO_CLIPBOARD',
-                  {text: selectedDirItemData.title}
-                )"
                 class="info-panel__header__title"
                 :class="{'cursor-pointer': inputState.ctrl}"
-              >{{selectedDirItemData.title}}
+                v-on="on"
+                @click.ctrl="$utils.copyToClipboard({
+                  text: itemTitle
+                })"
+              >
+                {{itemTitle}}
               </div>
             </template>
             <span>To copy: Ctrl + LClick</span>
           </v-tooltip>
 
           <!-- info-container::header::description -->
-          <div class="info-panel__header__sub-title">
-            {{selectedDirItemData.description}}
-          </div>
+          <div
+            class="info-panel__header__sub-title"
+            v-html="itemDescription"
+          />
         </v-layout>
 
-        <v-divider class="my-3"></v-divider>
+        <v-divider class="my-3" />
 
         <!-- info-container::properties -->
         <div class="info-panel__properties custom-scrollbar">
@@ -99,11 +86,7 @@ Copyright © 2021 - present Aleksey Hoffman. All rights reserved.
           >
             <!-- info-container::properties::size::title -->
             <div class="info-panel__properties__item__title">
-              {{
-                selectedDirItemData.type === 'directory'
-                  ? $localize.get('text_directory_size')
-                  : $localize.get('text_file_size')
-             }}:
+              {{itemSizeTitle}}:
             </div>
 
             <!-- button::get-size -->
@@ -111,20 +94,20 @@ Copyright © 2021 - present Aleksey Hoffman. All rights reserved.
               <div>
                 <v-btn
                   v-show="itemInfoSizeButton.show"
-                  @click.exact="itemInfoSizeButton.onClick"
                   class="info-panel__properties__item__value mb-0 button-2 fade-in-1s"
                   style="max-height: 20px;"
                   x-small
+                  @click.exact="handleItemInfoSizeButtonClick()"
                 >
                   <div
+                    v-show="showGetDirSizeBtn"
                     class="fade-in-1s"
-                    v-show="itemInfoSizeButton.state === 'idle'"
                   >
                     {{itemInfoSizeButton.title}}
                   </div>
                   <div
+                    v-show="showCancelGetDirSizeBtn"
                     class="fade-in-1s"
-                    v-show="itemInfoSizeButton.state === 'ongoing'"
                   >
                     <v-progress-circular
                       class="mr-2 fade-in-1s"
@@ -132,7 +115,7 @@ Copyright © 2021 - present Aleksey Hoffman. All rights reserved.
                       :color="$utils.getCSSVar('--color-6')"
                       size="12"
                       width="2"
-                    ></v-progress-circular>
+                    />
                     {{itemInfoSizeButton.title}}
                   </div>
                 </v-btn>
@@ -140,16 +123,15 @@ Copyright © 2021 - present Aleksey Hoffman. All rights reserved.
 
               <!-- property:size -->
               <v-tooltip bottom>
-                <template v-slot:activator="{ on }">
+                <template #activator="{on}">
                   <div
-                    v-on="on"
                     v-show="!showGetDirSizeBtn && !showCancelGetDirSizeBtn"
-                    @click.ctrl="$store.dispatch(
-                      'COPY_TEXT_TO_CLIPBOARD',
-                      {text: $utils.prettyBytes(selectedDirItems.getLast().stat.size, 1)}
-                    )"
                     class="info-panel__properties__item__value"
                     :class="{'cursor-pointer': inputState.ctrl}"
+                    v-on="on"
+                    @click.ctrl="$utils.copyToClipboard({
+                      text: $utils.prettyBytes(lastSelectedDirItem.stat.size, 1)
+                    })"
                   >
                     {{getSize}}
                   </div>
@@ -161,37 +143,47 @@ Copyright © 2021 - present Aleksey Hoffman. All rights reserved.
 
           <!-- properties -->
           <v-layout
-            v-for="(item, index) in selectedDirItemData.properties"
+            v-for="(item, index) in navigatorViewInfoPanel.data.properties"
             :key="index"
             class="info-panel__properties__item fade-in-500ms"
             column
           >
-            <!-- property-title -->
-            <div
-              class="info-panel__properties__item__title"
-              v-html="item.title"
-            >:
-            </div>
+            <template v-if="item.type === 'separator'">
+              <div class="info-panel__properties__item--separator">
+                <v-icon size="20px">
+                  mdi-information-outline
+                </v-icon>
+                {{item.title}}
+              </div>
+            </template>
 
-            <!-- property-value -->
-            <v-tooltip
-              :disabled="item.tooltip.length === 0"
-              bottom max-width="250px" offset-overflow
-            >
-              <template v-slot:activator="{ on }">
-                <div
-                  v-on="on"
-                  @click.ctrl="$store.dispatch(
-                    'COPY_TEXT_TO_CLIPBOARD',
-                    {text: item.value}
-                  )"
-                  class="info-panel__properties__item__value"
-                  :class="{'cursor-pointer': inputState.ctrl}"
-                >{{item.value}}
-                </div>
-              </template>
-              <span>{{item.tooltip}}</span>
-            </v-tooltip>
+            <template v-if="item.type !== 'separator'">
+              <div
+                class="info-panel__properties__item__title"
+                v-html="item.title"
+              >
+                :
+              </div>
+
+              <v-tooltip
+                :disabled="!item.tooltip || item.tooltip.length === 0"
+                bottom
+                max-width="400px"
+                offset-overflow
+              >
+                <template #activator="{on}">
+                  <div
+                    class="info-panel__properties__item__value"
+                    :class="{'cursor-pointer': inputState.ctrl}"
+                    v-on="on"
+                    @click.ctrl="handleClickPropertyValue({event: $event, item})"
+                  >
+                    {{item.value}}
+                  </div>
+                </template>
+                <span v-html="item.tooltip" />
+              </v-tooltip>
+            </template>
           </v-layout>
         </div>
       </div>
@@ -200,86 +192,105 @@ Copyright © 2021 - present Aleksey Hoffman. All rights reserved.
 </template>
 
 <script>
-import { mapGetters, mapState } from 'vuex'
-import { mapFields } from 'vuex-map-fields'
+import {mapGetters} from 'vuex'
+import {mapFields} from 'vuex-map-fields'
 
 export default {
   data () {
     return {
-      infoPanelAudioIsPlaying: false
+      infoPanelAudioIsPlaying: false,
+      infoPanelAudioDuration: 0,
+      itemInfoSizeButtonState: 'idle',
     }
   },
   created () {
-    this.delayInfoPanel()
+    this.showInfoPanel()
+  },
+  watch: {
+    'lastSelectedDirItem.path' () {
+      this.$nextTick(() => {
+        this.updateInfoPanelData()
+      })
+    },
   },
   computed: {
     ...mapGetters([
-      'selectedDirItems'
+      'selectedDirItems',
+      'lastSelectedDirItem',
     ]),
     ...mapFields({
-      navigatorViewInfoPanel: 'storageData.settings.infoPanels.navigatorView'
+      inputState: 'inputState',
+      navigatorViewInfoPanel: 'storageData.settings.infoPanels.navigatorView',
+      autoCalculateDirSize: 'storageData.settings.autoCalculateDirSize',
     }),
-    ...mapState({
-      selectedDirItemData: state => state.navigatorView.infoPanelData,
-      placeholders: state => state.placeholders,
-      inputState: state => state.inputState
-    }),
-    infoPanelThumbType () {
-      const itemMime = this.selectedDirItemData.mimeDescription
-      const itemIsDirectory = this.selectedDirItemData.type === 'directory'
-      const itemIsDirectorySymlink = this.selectedDirItemData.type === 'directory-symlink'
-      const itemIsFile = this.selectedDirItemData.type === 'file'
-      const itemIsFileSymlink = this.selectedDirItemData.type === 'file-symlink'
-      const itemMimeIsMediaType = ['image', 'video', 'audio'].includes(itemMime)
-      if (itemIsDirectory) {
-        return 'directory'
-      }
-      if (itemIsDirectorySymlink) {
-        return 'directory-symlink'
-      }
-      else if (itemIsFile && !itemMimeIsMediaType) {
-        return 'file'
-      }
-      else if (itemIsFileSymlink && !itemMimeIsMediaType) {
-        return 'file-symlink'
-      }
-      else if (itemMime === 'audio' && !this.infoPanelAudioIsPlaying) {
-        return 'file-audio'
-      }
-      else if (itemMime === 'audio' && this.infoPanelAudioIsPlaying) {
-        return 'file-audio-playing'
-      }
-      else {
-        return ''
-      }
+    itemExt () {
+      return this.$utils.getExt(this.lastSelectedDirItem?.path || '')
+    },
+    itemDescription () {
+      let type = this.lastSelectedDirItem?.type || ''
+      let dirItemCount = this.lastSelectedDirItem?.type === 'directory'
+        ? ` • ${this.lastSelectedDirItem?.dirItemCount} items`
+        : ''
+      let inaccessible = this.lastSelectedDirItem?.isInaccessible
+        ? ' • Inaccessible'
+        : ''
+      let isOffline = this.lastSelectedDirItem?.fsAttributes.isOffline
+        ? ' • Offline'
+        : ''
+      let keepOnDevice = this.lastSelectedDirItem?.fsAttributes.keepOnDevice
+        ? ' • Keep on device'
+        : ''
+      let itemMimeDescription = this.itemMimeDescription
+        ? ` • ${this.itemMimeDescription}`
+        : ''
+      let description = this.$localize.get(`text_${type.replace(/-/g, '_')}`)
+      return `${description}${dirItemCount}${itemMimeDescription}${isOffline}${keepOnDevice}${inaccessible}`
+    },
+    itemType () {
+      return this.lastSelectedDirItem?.type || ''
+    },
+    itemMimeDescription () {
+      return this.lastSelectedDirItem?.mime.mimeDescription || ''
+    },
+    itemTitle () {
+      return this.lastSelectedDirItem?.name || ''
+    },
+    itemSizeTitle () {
+      let type = this.lastSelectedDirItem?.type || ''
+      return type === 'directory'
+        ? this.$localize.get('text_directory_size')
+        : this.$localize.get('text_file_size')
     },
     previewIcon () {
+      const type = this.lastSelectedDirItem?.type
+      const itemMimeDescription = this.lastSelectedDirItem?.mime.mimeDescription
       const icon = {
         name: '',
         size: '56px',
-        class: 'info-panel__media-preview__icon'
+        class: 'info-panel__media-preview__icon',
       }
-      if (this.infoPanelThumbType === 'directory') {
+
+      if (type === 'directory') {
         icon.name = 'mdi-folder-outline'
       }
-      else if (this.infoPanelThumbType === 'directory-symlink') {
+      else if (type === 'directory-symlink') {
         icon.name = 'mdi-folder-move-outline'
       }
-      else if (this.infoPanelThumbType === 'file') {
-        icon.name = 'mdi-file-outline'
+      else if (type === 'file' && itemMimeDescription === 'image') {
+        icon.name = 'mdi-image-outline'
       }
-      else if (this.infoPanelThumbType === 'file-symlink') {
-        icon.name = 'mdi-file-move-outline'
-      }
-      else if (this.infoPanelThumbType === 'file-audio') {
+      else if (type === 'file' && itemMimeDescription === 'video') {
         icon.name = 'mdi-play-circle-outline'
+      }
+      else if (type === 'file' && itemMimeDescription === 'audio' && !this.infoPanelAudioIsPlaying) {
+        icon.name = 'mdi-play-outline'
         icon.size = '48px'
         icon.class = 'info-panel__media-preview__play info-panel__media-preview__icon'
         icon.ariaLabel = 'play pause toggle'
         icon.dataIcon = 'P'
         icon.onClick = this.toggleAudioPlay
       }
-      else if (this.infoPanelThumbType === 'file-audio-playing') {
+      else if (type === 'file' && itemMimeDescription === 'audio' && this.infoPanelAudioIsPlaying) {
         icon.name = 'mdi-pause-circle-outline'
         icon.size = '48px'
         icon.class = 'info-panel__media-preview__play info-panel__media-preview__icon'
@@ -287,69 +298,486 @@ export default {
         icon.dataIcon = 'P'
         icon.onClick = this.toggleAudioPlay
       }
+      else if (type === 'file') {
+        icon.name = 'mdi-file-outline'
+      }
+      else if (type === 'file-symlink') {
+        icon.name = 'mdi-file-move-outline'
+      }
+
       return icon
     },
-    previewMediaSrc () {
-      const itemMime = this.selectedDirItemData.mimeDescription
-      const mediaPath = this.selectedDirItems.getLast().realPath
-      if (itemMime === 'image') {
-        return this.$storeUtils.getSafePath(this.$utils.getUrlSafePath(mediaPath))
-      }
-      else if (itemMime === 'audio') {
-        return this.$storeUtils.getSafePath(this.$utils.getUrlSafePath(mediaPath))
-      }
-      else if (itemMime === 'video') {
-        return `${this.$storeUtils.getSafePath(this.$utils.getUrlSafePath(mediaPath))}#t=0.5`
-      }
-      else {
-        return ''
-      }
-    },
     showGetDirSizeBtn () {
-      return this.selectedDirItems?.getLast()?.stat?.size === null
+      return this.lastSelectedDirItem?.stat?.size === null && this.itemInfoSizeButtonState === 'idle'
     },
     showCancelGetDirSizeBtn () {
-      return this.selectedDirItems?.getLast()?.stat?.size === this.placeholders.calculatingDirSize
+      return this.lastSelectedDirItem?.stat?.size === null && this.itemInfoSizeButtonState === 'ongoing'
     },
     itemInfoSizeButton () {
       const show = this.showGetDirSizeBtn || this.showCancelGetDirSizeBtn
       if (this.showCancelGetDirSizeBtn) {
         return {
           show,
-          title: 'cancel',
-          state: 'ongoing',
-          onClick: () => this.$store.dispatch('CANCEL_FETCH_CURRENT_DIR_SIZE')
+          title: this.$localize.get('text_cancel'),
         }
       }
       else {
         return {
           show,
           title: this.$localize.get('text_get_size'),
-          state: 'idle',
-          onClick: () => this.$store.dispatch('FETCH_CURRENT_DIR_SIZE')
         }
       }
     },
     getSize () {
       try {
-        return this.$utils.prettyBytes(this.selectedDirItems.getLast().stat.size, 1)
+        const sizeOnDisk = this.lastSelectedDirItem?.sizeOnDisk === 0
+          ? ' • 0 bytes on disk'
+          : ''
+        const size = this.$utils.prettyBytes(this.lastSelectedDirItem?.stat?.size, 1)
+        return `${size}${sizeOnDisk}`
       }
       catch (error) {
         return 'unknown'
       }
-    }
+    },
+    shouldFetchMediaPreview () {
+      const isInaccessible = this.lastSelectedDirItem?.isInaccessible
+      const itemIsOffline = this.lastSelectedDirItem?.fsAttributes.isOffline
+      const isMediaType = this.lastSelectedDirItem.isMediaType
+      return (!isInaccessible && !itemIsOffline) || !isMediaType
+    },
+    formattedInfoPanelAudioDuration () {
+      return this.$utils.getFormattedTime({
+        time: this.infoPanelAudioDuration,
+        unit: 'seconds',
+        format: 'HH:mm:ss',
+      })
+    },
   },
   methods: {
-    showPreviewItem (mediaType) {
-      const itemMime = this.selectedDirItemData.mimeDescription
-      return itemMime === mediaType
+    async handleItemInfoSizeButtonClick () {
+      if (this.itemInfoSizeButtonState === 'idle') {
+        this.itemInfoSizeButtonState = 'ongoing'
+        await this.$store.dispatch('FETCH_CURRENT_DIR_SIZE')
+        this.itemInfoSizeButtonState = 'idle'
+      }
+      else {
+        this.itemInfoSizeButtonState = 'idle'
+        this.$store.dispatch('TERMINATE_ALL_FETCH_DIR_SIZE')
+      }
     },
-    delayInfoPanel () {
-      if (this.navigatorViewInfoPanel) {
-        this.navigatorViewInfoPanel = false
+    async updateInfoPanelData () {
+      this.navigatorViewInfoPanel.data = this.getInfoPanelData()
+      await this.fetchPreview()
+      this.cancelFetchDirSize()
+      this.initAutoFetchDirSize()
+    },
+    async initAutoFetchDirSize () {
+      if (this.autoCalculateDirSize) {
+        if (this.itemInfoSizeButtonState === 'idle') {
+          this.itemInfoSizeButtonState = 'ongoing'
+          await this.$store.dispatch('AUTO_CALCULATE_DIR_SIZE', this.lastSelectedDirItem)
+          this.itemInfoSizeButtonState = 'idle'
+        }
+      }
+    },
+    cancelFetchDirSize () {
+      this.itemInfoSizeButtonState = 'idle'
+      this.$store.dispatch('TERMINATE_ALL_FETCH_DIR_SIZE')
+    },
+    getPreviewMediaSrc () {
+      const itemMimeDescription = this.lastSelectedDirItem?.mime.mimeDescription
+      const mediaPath = this.lastSelectedDirItem?.realPath || ''
+      if (itemMimeDescription === 'image') {
+        return this.$storeUtils.getSafePath(this.$sharedUtils.getUrlSafePath(mediaPath))
+      }
+      else if (itemMimeDescription === 'audio') {
+        return this.$storeUtils.getSafePath(this.$sharedUtils.getUrlSafePath(mediaPath))
+      }
+      else if (itemMimeDescription === 'video') {
+        return `${this.$storeUtils.getSafePath(this.$sharedUtils.getUrlSafePath(mediaPath))}#t=0.5`
+      }
+      else {
+        return ''
+      }
+    },
+    getInfoPanelData () {
+      let item = this.lastSelectedDirItem
+      if (!item.path) {return}
+      const {isReadOnly, permissions} = this.$utils.getItemPermissions(item)
+      const isDirItemPinned = this.$store.state.storageData.pinned.items
+        .some(listItem => listItem.path === item.path)
+      const isDirItemProtected = this.$store.state.storageData.protected.items
+        .some(listItem => listItem.path === item.path)
+      const copyShortcut = 'Ctrl + LClick'
+      const copyWithSecondSlashShortcut = 'Ctrl + Alt + LClick'
+      const copyWithQuotesShortcut = 'Ctrl + Shift + LClick'
+
+      let copyPathTooltip
+      if (this.$sharedUtils.platform === 'win32') {
+        copyPathTooltip = `
+          <div>
+            <span class="inline-code--light">${copyShortcut}</span> 
+              - copy path
+          </div>
+          <div>
+            <span class="inline-code--light">${copyWithSecondSlashShortcut}</span> 
+              - copy path | add second slash 
+          </div>
+          <div>
+            <span class="inline-code--light">${copyWithQuotesShortcut}</span> 
+              - copy path | wrap with quotes
+          </div>
+        `
+      }
+      else {
+        copyPathTooltip = `
+          <div>
+            <span class="inline-code--light">${copyShortcut}</span> 
+              - copy path
+          </div>
+          <div>
+            <span class="inline-code--light">${copyWithQuotesShortcut}</span> 
+              - copy path | wrap with quotes
+          </div>
+        `
+      }
+
+      // Define main properties
+      let properties = [
+        {
+          propName: 'path',
+          title: this.$localize.get('text_path'),
+          value: item.path,
+          tooltip: copyPathTooltip,
+        },
+        {
+          propName: 'dateCreated',
+          title: this.$localize.get('text_created'),
+          value: this.$utils.getLocalDateTime(item.stat.birthtime, this.$store.state.storageData.settings.dateTime),
+          tooltip: `${this.$localize.get('tooltip_text_to_copy')}: ${copyShortcut}`,
+        },
+        {
+          propName: 'dateModified',
+          title: this.$localize.get('text_modified'),
+          value: this.$utils.getLocalDateTime(item.stat.mtime, this.$store.state.storageData.settings.dateTime),
+          tooltip: `${this.$localize.get('tooltip_text_to_copy')}: ${copyShortcut}`,
+        },
+        {
+          propName: 'dateChanged',
+          title: this.$localize.get('text_changed'),
+          value: this.$utils.getLocalDateTime(item.stat.ctime, this.$store.state.storageData.settings.dateTime),
+          tooltip: `${this.$localize.get('tooltip_text_to_copy')}: ${copyShortcut}`,
+        },
+        {
+          propName: 'permissions',
+          title: this.$localize.get('text_mode'),
+          value: permissions,
+          tooltip: `${this.$localize.get('tooltip_text_to_copy')}: ${copyShortcut}`,
+        },
+        {
+          propName: 'protected',
+          title: this.$localize.get('text_protected'),
+          value: isDirItemProtected
+            ? this.$localize.get('text_yes')
+            : this.$localize.get('text_no'),
+          tooltip: `
+            Protected items cannot be modified or deleted (from within this app only). 
+            Protected items can be found on the dashboard page.
+          `,
+        },
+        {
+          propName: 'pinned',
+          title: this.$localize.get('text_pinned'),
+          value: isDirItemPinned
+            ? this.$localize.get('text_yes')
+            : this.$localize.get('text_no'),
+          tooltip: 'Pinned items can be found on the dashboard page',
+        },
+      ]
+
+      // Append conditional properties
+      const realPath = {
+        propName: 'realPath',
+        title: this.$localize.get('text_real_path'),
+        value: item.realPath,
+        tooltip: copyPathTooltip,
+      }
+      if (item.path !== item.realPath) {
+        properties.splice(1, 0, realPath)
+      }
+
+      // Append media info properties
+      if (this.shouldFetchMediaPreview) {
+        this.fetchMediaInfo()
+          .then((data) => {
+            if (data.length > 0) {
+              properties.push({
+                type: 'separator',
+                title: this.$localize.get('media_info'),
+              })
+            }
+
+            data.forEach(prop => {
+              properties.push(prop)
+            })
+          })
+      }
+
+      return {
+        properties,
+      }
+    },
+    async fetchPreview () {
+      const isMediaType = this.lastSelectedDirItem.isMediaType
+      this.clearMediaContainer()
+      if (!this.shouldFetchMediaPreview) {return}
+
+      if (isMediaType) {
+        this.fetchMediaPreview()
+      }
+      else {
+        this.fetchNonMediaPreview()
+      }
+    },
+    formatValue (params) {
+      if (params.property === 'duration') {
+        return this.$utils.getFormattedTime({
+          time: params.propertyValue,
+          unit: 'seconds',
+          format: 'HH:mm:ss',
+        })
+      }
+      else if (params.property === 'avg_frame_rate') {
+        if (typeof params.propertyValue === 'string') {
+          let splitValue = params.propertyValue.split('/')
+          return Number(`${Number(splitValue[0]) / Number(splitValue[1])}`).toFixed(3)
+        }
+        else {
+          return params.propertyValue
+        }
+      }
+      else if (params.property === 'bit_rate') {
+        return `${this.$utils.prettyBytes(Number(params.propertyValue), 1)} / ${this.$localize.get('sec', {capitalize: false})}`
+      }
+      else {
+        return params.propertyValue
+      }
+    },
+    formatMediaInfo (data) {
+      if (!data || !this.lastSelectedDirItem) {return []}
+
+      const itemMimeDescription = this.lastSelectedDirItem.mime.mimeDescription
+      let dataFormatted = []
+
+      if (itemMimeDescription === 'image') {
+        let displayedImageStreamProperties = [
+          'width',
+          'height',
+          'bits_per_raw_sample',
+          'color_space',
+          'pix_fmt',
+          'chroma_location',
+        ]
+        displayedImageStreamProperties.forEach(property => {
+          if (data.streams[0]) {
+            dataFormatted.push({
+              propName: property,
+              title: this.$localize.get(property),
+              value: this.formatValue({property, propertyValue: data.streams[0][property]}),
+            })
+          }
+        })
+      }
+      else if (itemMimeDescription === 'audio') {
+        let displayedAudioFormatProperties = [
+          'duration',
+          'bit_rate',
+        ]
+        let displayedAudioFormatTagsProperties = [
+          'encoder',
+        ]
+        let displayedAudioStreamProperties = [
+          'channel_layout',
+          'channels',
+          'sample_rate',
+        ]
+        let displayedAudioStreamTagsProperties = [
+          'encoder',
+        ]
+        displayedAudioFormatProperties.forEach(property => {
+          if (data.format) {
+            dataFormatted.push({
+              propName: property,
+              title: this.$localize.get(property),
+              value: this.formatValue({property, propertyValue: data.format[property]}),
+            })
+          }
+        })
+        displayedAudioFormatTagsProperties.forEach(property => {
+          if (data.format.tags) {
+            dataFormatted.push({
+              propName: property,
+              title: this.$localize.get(property),
+              value: this.formatValue({property, propertyValue: data.format.tags[property]}),
+            })
+          }
+        })
+        data.streams.forEach((stream, index) => {
+          displayedAudioStreamProperties.forEach(property => {
+            dataFormatted.push({
+              propName: property,
+              title: `${this.$localize.get('stream')} ${index + 1} • ${this.$localize.get(property)}`,
+              value: this.formatValue({property, propertyValue: stream[property]}),
+            })
+          })
+          displayedAudioStreamTagsProperties.forEach(property => {
+            dataFormatted.push({
+              propName: property,
+              title: `${this.$localize.get('stream')} ${index + 1} • ${this.$localize.get(property)}`,
+              value: this.formatValue({property, propertyValue: stream.tags[property]}),
+            })
+          })
+        })
+      }
+      else if (itemMimeDescription === 'video') {
+        let displayedVideoFormatProperties = [
+          'duration',
+          'bit_rate',
+        ]
+        let displayedVideoFormatTagsProperties = [
+          'encoder',
+        ]
+        let displayedVideoStreamProperties = [
+          'codec_name',
+          'codec_tag_string',
+          'width',
+          'height',
+          'display_aspect_ratio',
+          'avg_frame_rate',
+          'bits_per_raw_sample',
+          'bit_rate',
+          'chroma_location',
+          'has_b_frames',
+          'pix_fmt',
+          'profile',
+          'closed_captions',
+          'channel_layout',
+          'channels',
+          'sample_rate',
+        ]
+        let displayedVideoStreamTagsProperties = [
+          'encoder',
+        ]
+        displayedVideoFormatProperties.forEach(property => {
+          if (data.format) {
+            dataFormatted.push({
+              propName: property,
+              title: this.$localize.get(property),
+              value: this.formatValue({property, propertyValue: data.format[property]}),
+            })
+          }
+        })
+        displayedVideoFormatTagsProperties.forEach(property => {
+          if (data.format.tags) {
+            dataFormatted.push({
+              propName: property,
+              title: this.$localize.get(property),
+              value: this.formatValue({property, propertyValue: data.format.tags[property]}),
+            })
+          }
+        })
+        data.streams.forEach((stream, index) => {
+          displayedVideoStreamProperties.forEach(property => {
+            dataFormatted.push({
+              propName: property,
+              title: `${this.$localize.get('stream')} ${index + 1} • ${this.$localize.get(property)}`,
+              value: this.formatValue({property, propertyValue: stream[property]}),
+            })
+          })
+          displayedVideoStreamTagsProperties.forEach(property => {
+            dataFormatted.push({
+              propName: property,
+              title: `${this.$localize.get('stream')} ${index + 1} • ${this.$localize.get(property)}`,
+              value: this.formatValue({property, propertyValue: stream.tags[property]}),
+            })
+          })
+        })
+      }
+
+      return dataFormatted.filter(item => item.value !== undefined)
+    },
+    async fetchMediaInfo () {
+      return new Promise((resolve, reject) => {
+        this.$store.dispatch('GET_MEDIA_FILE_INFO', {
+          path: this.lastSelectedDirItem.realPath,
+        })
+          .then(event => resolve(this.formatMediaInfo(event.data.metadata)))
+      })
+    },
+    async fetchMediaPreview () {
+      const itemMimeDescription = this.lastSelectedDirItem?.mime.mimeDescription
+      let mediaNode
+      if (itemMimeDescription === 'image') {
+        mediaNode = new Image()
+        await this.setImageMedia(mediaNode)
+        // this.$store.dispatch('GENERATE_THUMBNAIL', {item: this.lastSelectedDirItem})
+      }
+      else if (itemMimeDescription === 'audio') {
+        mediaNode = new Audio()
+        await this.setAudioMedia(mediaNode)
+        this.infoPanelAudioDuration = 0
+        mediaNode.onloadedmetadata = () => {
+          this.infoPanelAudioDuration = mediaNode.duration
+        }
+      }
+      else if (itemMimeDescription === 'video') {
+        mediaNode = document.createElement('video')
+        await this.setVideoMedia(mediaNode)
+      }
+      mediaNode.classList.add('info-panel__preview-container__media', 'fade-in-1s')
+      this.replaceMedia(mediaNode)
+    },
+    fetchNonMediaPreview () {
+      this.clearMediaContainer()
+    },
+    clearMediaContainer () {
+      this.replaceMedia(document.createElement('div'))
+    },
+    async setImageMedia (mediaNode) {
+      mediaNode.src = this.getPreviewMediaSrc()
+      await mediaNode.decode()
+    },
+    async setAudioMedia (mediaNode) {
+      mediaNode.src = this.getPreviewMediaSrc()
+      mediaNode.setAttribute('controls', true)
+      mediaNode.setAttribute('preload', 'metadata')
+      mediaNode.style.display = 'none'
+    },
+    async setVideoMedia (mediaNode) {
+      mediaNode.src = this.getPreviewMediaSrc()
+      mediaNode.setAttribute('controls', true)
+      mediaNode.setAttribute('preload', 'metadata')
+    },
+    replaceMedia (mediaNode) {
+      let mediaContainerNode = document.querySelector('.info-panel__preview-container__media-container')
+      mediaContainerNode.replaceChildren(mediaNode)
+    },
+    handleClickPropertyValue (params) {
+      const asPath = ['path', 'realPath'].includes(params.item.propName)
+
+      this.$utils.copyToClipboard({
+        text: params.item.value,
+        asPath,
+        event: params.event,
+      })
+    },
+    showInfoPanel () {
+      if (this.navigatorViewInfoPanel.value) {
+        this.navigatorViewInfoPanel.value = false
         setTimeout(() => {
-          this.navigatorViewInfoPanel = true
-        }, 100)
+          this.navigatorViewInfoPanel.value = true
+        }, 200)
       }
     },
     toggleAudioPlay () {
@@ -365,8 +793,8 @@ export default {
         this.infoPanelAudioIsPlaying = false
         media.pause()
       }
-    }
-  }
+    },
+  },
 }
 </script>
 
@@ -381,23 +809,28 @@ export default {
 
 .info-panel
   .v-divider {
-    background-color: var(--divider-color-2) !important;
+    background-color: var(--divider-color-4) !important;
   }
 
 .info-panel__preview-container {
+  position: relative;
   width: 280px;
   height: 158px;
-  background-color: var(--info-panel-preview-container-bg-color);
-  position: relative;
+  background-color: rgba(255, 255, 255, 0.02);
 }
 
+#app[dir-item-background="none"]
+  .info-panel__preview-container {
+    background-color: transparent;
+  }
+
 .info-panel__preview-container__media-container * {
+  position: absolute;
   width: 100%;
   height: 100%;
   top: 0;
   left: 0;
   object-fit: cover;
-  position: absolute;
 }
 
 .info-panel__preview-container__media-container *:focus {
@@ -405,9 +838,26 @@ export default {
   border: none;
 }
 
+.info-panel__preview-container__info-overlay {
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  top: 0;
+  left: 0;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+}
+
+.info-panel__preview-container__info-overlay__audio-duration {
+  margin-top: 6px;
+  font-size: 14px;
+}
+
 .info-panel__media-preview__play {
-  cursor: pointer;
   z-index: 1;
+  cursor: pointer;
 }
 
 .info-panel__info-container {
@@ -416,9 +866,9 @@ export default {
 }
 
 .info-panel__header {
-  margin-top: 8px;
   display: flex;
   align-items: center;
+  margin-top: 8px;
 }
 
 .info-panel__header
@@ -430,18 +880,17 @@ export default {
   }
 
 .info-panel__header__title {
-  font-size: 18px;
   color: var(--title-color-1);
+  font-size: 18px;
 }
 
 .info-panel__header__sub-title {
-  font-size: 14px;
   color: var(--title-color-1);
+  font-size: 14px;
 }
 
 .info-panel__properties {
   position: relative;
-  color: var(--color-1);
   margin-bottom: 8px;
   padding-bottom: 16px;
   height: calc(
@@ -450,6 +899,7 @@ export default {
     - var(--action-toolbar-height)
     - 158px - 32px - 16px - 36px
   );
+  color: var(--color-1);
   -webkit-mask-image: linear-gradient(180deg, #fff 85%,transparent);
 }
 
@@ -463,19 +913,38 @@ export default {
     height: 24px !important;
   }
 
+.info-panel__properties__item--separator {
+  display: flex;
+  align-items: center;
+  padding: 8px 4px;
+  color: var(--color-7);
+  border-top: 1px solid var(--highlight-color-4);
+  border-bottom: 1px solid var(--highlight-color-4);
+  border-radius: 4px;
+  font-size: 14px;
+  text-transform: uppercase;
+  user-select: none;
+}
+
+.info-panel__properties__item--separator
+  .v-icon {
+    margin-right: 12px;
+    color: var(--color-7) !important;
+  }
+
 .info-panel__properties__item__title {
   user-select: none;
+  margin-right: 8px;
   color: var(--color-6);
   font-size: 12px;
-  margin-right: 8px;
-  /* text-transform: capitalize; */
   text-transform: uppercase;
 }
 
 .info-panel__properties__item__value {
+  width: fit-content;
+  margin-bottom: 8px;
   color: var(--color-7);
   font-size: 14px;
-  margin-bottom: 8px;
   word-break: break-all;
 }
 
