@@ -666,7 +666,10 @@ NOTES:
         className: 'os-theme-minimal-light',
         scrollbars: { autoHide: 'move' }
       }"
-      class="sticky-scroller__condtent text-editor__content-container fade-mask--bottom--10"
+      class="sticky-scroller__condtent text-editor__content-container fade-mask--bottom"
+      :style="{
+        '--fade-mask-bottom': '15%'
+      }"
     >
       <span @click.right.exact="showNoteEditorContextMenu">
         <div v-show="showSourceCode" class="text-editor__source-code-container">
@@ -686,20 +689,21 @@ NOTES:
           :spellcheck="spellcheck"
           :isEmpty="isEmpty"
           :isReadOnly="readOnly"
-        ><iframe width="560" height="315" src="https://www.youtube.com/embed/HfvJVcGeMmg" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe></div>
+        ></div>
       </span>
     </overlay-scrollbars>
   </div>
 </template>
 
 <script>
-import { mapFields } from 'vuex-map-fields'
+import {mapFields} from 'vuex-map-fields'
+import {colors} from '../utils/colors.js'
 import Filehasher from '../utils/fileHasher.js'
 import TimeUtils from '../utils/timeUtils.js'
 import katex from 'katex'
 import 'katex/dist/katex.min.css'
-import { colors } from '../utils/colors.js'
-const { ipcRenderer } = require('electron')
+
+const {ipcRenderer} = require('electron')
 const fs = require('fs')
 const PATH = require('path')
 
@@ -751,7 +755,7 @@ export default {
       templates: [
         {
           title: 'template 1',
-          content: '<b>Template 1 content:</b> test'
+          content: '<strong>Template 1 content:</strong> test'
         },
         {
           title: 'template 2',
@@ -816,8 +820,8 @@ export default {
   },
   computed: {
     ...mapFields({
-      appPaths: 'appPaths',
-      dialogs: 'dialogs'
+      appPaths: 'storageData.settings.appPaths',
+      dialogs: 'dialogs',
     }),
     isEmpty () {
       return this.charCount === null || this.charCount === 0
@@ -1634,6 +1638,7 @@ export default {
   rgb(255, 255, 255) and caret-color when you delete some text in-between strings */
   background-color: var(--bg-color-1);
   caret-color: var(--color-6);
+  font-size: 16px !important;
 }
 
 .text-editor__content:focus {
@@ -1701,10 +1706,6 @@ export default {
   input[type="checkbox"]:checked + label {
     color: var(--color-6);
   }
-
-.text-editor__content {
-  font-size: 16px !important;
-}
 
 .text-editor__content
   code:after,
