@@ -7,15 +7,22 @@ Copyright © 2021 - present Aleksey Hoffman. All rights reserved.
   <v-app-bar
     id="window-toolbar"
     :style="{'background': windowToolbarBackgroundColor}"
-    app flat clipped-right clipped-left height="32px"
+    app
+    flat
+    clipped-right
+    clipped-left
+    height="32px"
   >
+    <div class="window-toolbar__drag-region" />
     <div class="window-toolbar__content-container">
       <div
-        class="window-toolbar__header-container"
-        @click="navigationPanelMiniVariant = !navigationPanelMiniVariant"
-        :style="!navigationPanelMiniVariant ? 'width: 278px' : 'width: 64px'"
         v-ripple
+        class="window-toolbar__header-container"
+        :style="!navigationPanelMiniVariant
+          ? 'width: var(--nav-panel-width-expanded)'
+          : 'width: var(--nav-panel-width)'"
         align-center
+        @click="toggleNavigationPanel()"
       >
         <div class="nav-panel__item__icon-container pl-1">
           <v-icon class="nav-panel__item__icon window-toolbar__item">
@@ -25,14 +32,18 @@ Copyright © 2021 - present Aleksey Hoffman. All rights reserved.
         <transition name="slide-fade-left">
           <div v-if="!navigationPanelMiniVariant">
             <div class="nav-panel__item__title">
-              Sigma file manager
+              Sigma File Manager
             </div>
           </div>
         </transition>
       </div>
 
-      <v-layout class="window-toolbar__content--main">
-
+      <v-layout
+        class="window-toolbar__content--main"
+        :style="{
+          'grid-template-columns': gridTemplateColumnsStyle
+        }"
+      >
         <!-- item-iterator -->
         <div
           v-for="(item, index) in windowToolbar.items"
@@ -42,22 +53,32 @@ Copyright © 2021 - present Aleksey Hoffman. All rights reserved.
           }"
         >
           <template v-if="item.type === 'workspaces-menu'">
-            <workspaces-menu :iconColor="windowToolbarFontColor"/>
+            <workspaces-menu :icon-color="windowToolbarFontColor" />
           </template>
+
           <template v-if="item.type === 'tabs-menu'">
-            <tabs-menu :iconColor="windowToolbarFontColor"/>
+            <tabs-menu :icon-color="windowToolbarFontColor" />
           </template>
-          <template v-if="item.type === 'notification-menu'">
-            <notification-menu :iconColor="windowToolbarFontColor"/>
+
+          <template v-if="showNavigatorTabBar(item) && item.breakpoint && $vuetify.breakpoint[item.breakpoint]">
+            <navigator-tab-bar />
           </template>
+
+          <template v-if="item.type === 'notification-menu' && item.ifCondition">
+            <notification-menu :icon-color="windowToolbarFontColor" />
+          </template>
+
           <template v-if="item.type === 'menu-button'">
-            <v-tooltip bottom>
-              <template v-slot:activator="{ on }">
+            <v-tooltip
+              bottom
+              min-width="100px"
+            >
+              <template #activator="{ on }">
                 <v-btn
                   class="window-toolbar__item"
+                  icon
                   v-on="on"
                   @click="item.onClick()"
-                  icon
                 >
                   <v-icon
                     :color="windowToolbarFontColor"
@@ -71,14 +92,55 @@ Copyright © 2021 - present Aleksey Hoffman. All rights reserved.
                 <div
                   class="tooltip__description"
                   v-html="item.tooltip.description"
-                ></div>
+                />
                 <div
-                  class="tooltip__shortcut"
                   v-show="[undefined, true].includes(item.tooltip.shortutIfCondition)"
+                  class="tooltip__shortcut"
                   v-html="item.tooltip.shortcut"
-                ></div>
+                />
               </span>
             </v-tooltip>
+          </template>
+
+          <template v-if="item.type === 'window-controls-group'">
+            <div class="window-controls-group">
+              <div
+                v-for="(item, index) in item.items"
+                :key="'window-controls-group-item-' + index"
+              >
+                <v-tooltip
+                  bottom
+                  min-width="100px"
+                >
+                  <template #activator="{ on }">
+                    <button
+                      class="window-toolbar__item window-controls-group__item"
+                      icon
+                      v-on="on"
+                      @click="item.onClick()"
+                    >
+                      <v-icon
+                        :color="windowToolbarFontColor"
+                        :size="item.icon.size"
+                      >
+                        {{item.icon.name}}
+                      </v-icon>
+                    </button>
+                  </template>
+                  <span>
+                    <div
+                      class="tooltip__description"
+                      v-html="item.tooltip.description"
+                    />
+                    <div
+                      v-show="[undefined, true].includes(item.tooltip.shortutIfCondition)"
+                      class="tooltip__shortcut"
+                      v-html="item.tooltip.shortcut"
+                    />
+                  </span>
+                </v-tooltip>
+              </div>
+            </div>
           </template>
         </div>
       </v-layout>
@@ -88,20 +150,23 @@ Copyright © 2021 - present Aleksey Hoffman. All rights reserved.
 
 <script>
 import electron from 'electron'
-import { mapFields } from 'vuex-map-fields'
-import { mapGetters } from 'vuex'
+import {mapFields} from 'vuex-map-fields'
+import {mapGetters} from 'vuex'
 const electronRemote = require('@electron/remote')
 
 export default {
   data () {
     return {
-      window: electronRemote.getCurrentWindow()
+      window: electronRemote.getCurrentWindow(),
+      animationTiming: {
+        navigationPanel: 500,
+      },
     }
   },
   computed: {
     ...mapGetters([
       'systemInfo',
-      'windowToolbarFontColor'
+      'windowToolbarFontColor',
     ]),
     ...mapFields({
       shortcuts: 'storageData.settings.shortcuts',
@@ -110,9 +175,11 @@ export default {
       navigationPanelMiniVariant: 'navigationPanel.miniVariant',
       homeBannerValue: 'storageData.settings.homeBanner.value',
       windowCloseButtonAction: 'storageData.settings.windowCloseButtonAction',
-      homeBannerOverlaySelectedItem: 'storageData.settings.homeBanner.overlay.selectedItem'
+      homeBannerOverlaySelectedItem: 'storageData.settings.homeBanner.overlay.selectedItem',
+      navigatorTabLayout: 'storageData.settings.navigator.tabs.layout',
+      notifications: 'notifications',
     }),
-    toolbarColor () { return this.$store.state.storageData.settings.theme.toolbarColor },
+    toolbarColor () {return this.$store.state.storageData.settings.theme.toolbarColor},
     windowToolbarFontColor () {
       // Force white icons on transparent toolbars
       if (this.toolbarIsTransparent) {
@@ -151,43 +218,64 @@ export default {
         return 'Close the app'
       }
     },
+    hiddenNotifications () {
+      return [
+        ...this.notifications.filter(item => item.isHidden && item.isPinned),
+        ...this.notifications.filter(item => item.isHidden && !item.isPinned),
+      ]
+    },
+    gridTemplateColumnsStyle () {
+      return this.windowToolbar.items.map(item => {
+        if (item.ifCondition === false) {
+          return '0'
+        }
+
+        if (item.breakpoint) {
+          return this.$vuetify.breakpoint[item.breakpoint]
+            ? item.width
+            : '0'
+        }
+        else {
+          return item.width
+        }
+      }).join(' ')
+    },
     windowToolbar () {
       return {
         items: [
           {
-            type: 'workspaces-menu'
-          },
-          {
-            type: 'tabs-menu'
-          },
-          {
             type: 'menu-button',
-            icon: {
-              size: '20px',
-              name: 'mdi-pound'
-            },
-            tooltip: {
-              description: 'Shortcut list',
-              shortcut: this.shortcuts.shortcutsDialog.shortcut
-            },
-            onClick: () => { this.dialogs.shortcutsDialog.value = true }
-          },
-          {
-            type: 'menu-button',
+            width: '32px',
             icon: {
               size: '18px',
-              name: 'mdi-comment-question-outline'
+              name: 'mdi-message-question-outline',
             },
             tooltip: {
-              description: 'App guide'
+              description: 'App guide',
             },
-            onClick: () => { this.dialogs.guideDialog.value = true }
+            onClick: () => {this.dialogs.guideDialog.value = true},
           },
           {
-            type: 'notification-menu'
+            type: 'workspaces-menu',
+            width: '48px',
           },
           {
-            type: 'spacer'
+            type: 'tabs-menu',
+            width: '48px',
+          },
+          {
+            type: 'navigator-tab-bar',
+            width: 'minmax(300px, 50%)',
+            breakpoint: 'mdAndUp',
+          },
+          {
+            type: 'spacer',
+            width: 'minmax(64px, 1fr)',
+          },
+          {
+            type: 'notification-menu',
+            ifCondition: this.hiddenNotifications.length > 0,
+            width: '32px',
           },
           // TODO: finish in v1.X
           // {
@@ -213,64 +301,92 @@ export default {
           //   onClick: () => undefined
           // },
           {
-            type: 'menu-button',
-            icon: {
-              size: '26px',
-              name: 'mdi-chevron-up'
-            },
-            tooltip: {
-              description: 'Scroll top',
-              shortcut: this.shortcuts.scrollTop.shortcut
-            },
-            onClick: () => this.$store.dispatch('SCROLL_TOP_CONTENT_AREA')
+            type: 'window-controls-group',
+            width: '128px',
+            items: [
+              {
+                type: 'window-controls',
+                icon: {
+                  size: '18px',
+                  name: 'mdi-minus',
+                },
+                tooltip: {
+                  description: 'Minimize window',
+                  shortcut: this.shortcuts.toggleApp.shortcut,
+                },
+                onClick: () => this.minimizeWindow(),
+              },
+              {
+                type: 'window-controls',
+                icon: {
+                  size: '16px',
+                  name: 'mdi-aspect-ratio',
+                },
+                tooltip: {
+                  description: 'Toggle window size',
+                  shortcut: this.shortcuts.windowPosition.shortcut[this.systemInfo.platform],
+                },
+                onClick: () => this.maximizeWindow(),
+              },
+              {
+                type: 'window-controls',
+                icon: {
+                  size: '18px',
+                  name: 'mdi-close',
+                },
+                tooltip: {
+                  description: `
+                    Close window:
+                    <br>Action: ${this.windowCloseButtonActionDescription}
+                  `,
+                  shortcut: `
+                    Toggle window: ${this.shortcuts.toggleApp.shortcut}
+                  `,
+                  shortutIfCondition: this.windowCloseButtonAction !== 'closeApp',
+                },
+                onClick: () => this.closeWindow(),
+              },
+            ],
           },
-          {
-            type: 'menu-button',
-            icon: {
-              size: '18px',
-              name: 'mdi-minus'
-            },
-            tooltip: {
-              description: 'Minimize window',
-              shortcut: this.shortcuts.toggleApp.shortcut
-            },
-            onClick: () => this.minimizeWindow()
-          },
-          {
-            type: 'menu-button',
-            icon: {
-              size: '16px',
-              name: 'mdi-aspect-ratio'
-            },
-            tooltip: {
-              description: 'Toggle window size',
-              shortcut: this.shortcuts.windowPosition.shortcut[this.systemInfo.platform]
-            },
-            onClick: () => this.maximizeWindow()
-          },
-          {
-            type: 'menu-button',
-            icon: {
-              size: '18px',
-              name: 'mdi-close'
-            },
-            tooltip: {
-              description: `
-                Close window:
-                <br>Action: ${this.windowCloseButtonActionDescription}
-              `,
-              shortcut: `
-                Toggle window: ${this.shortcuts.toggleApp.shortcut}
-              `,
-              shortutIfCondition: this.windowCloseButtonAction !== 'closeApp'
-            },
-            onClick: () => this.closeWindow()
-          }
-        ]
+        ],
       }
-    }
+    },
   },
   methods: {
+    showNavigatorTabBar (item) {
+      return item.type === 'navigator-tab-bar' &&
+      this.$route.name === 'navigator' &&
+      this.navigatorTabLayout === 'compact-vertical-and-traditional-horizontal'
+    },
+    toggleNavigationPanel () {
+      this.toggleShadow()
+      this.$nextTick(() => {
+        this.navigationPanelMiniVariant = !this.navigationPanelMiniVariant
+      })
+    },
+    /** This function optimizes rendering performance (20FPS => 60FPS)
+    * by disabling shadows during layout transition
+    */
+    toggleShadow () {
+      this.removeShadow()
+      setTimeout(() => {
+        this.addShadow()
+      }, this.animationTiming.navigationPanel)
+    },
+    removeShadow () {
+      let cards = document.querySelectorAll('.item-card')
+      cards.forEach(card => {
+        card.setAttribute('optimized', true)
+        card.setAttribute('optimized', true)
+      })
+    },
+    addShadow () {
+      let cards = document.querySelectorAll('.item-card')
+      cards.forEach(card => {
+        card.removeAttribute('optimized')
+        card.removeAttribute('optimized')
+      })
+    },
     minimizeWindow () {
       this.window.minimize()
     },
@@ -284,8 +400,8 @@ export default {
     },
     closeWindow () {
       electron.ipcRenderer.send('handle:close-app', this.windowCloseButtonAction)
-    }
-  }
+    },
+  },
 }
 </script>
 
@@ -293,26 +409,59 @@ export default {
 #window-toolbar {
   transition: background-color 0.1s;
   z-index: 10;
+}
+
+#window-toolbar
+  .window-toolbar__content--main {
+    display: flex;
+    align-items: center;
+    gap: var(--toolbar-item-gap);
+    padding:
+      0px
+      var(--window-toolbar-padding-right)
+      0px
+      var(--window-toolbar-padding-left) !important;
+    display: grid;
+    width: calc(100vw - 301px);
+  }
+
+.window-toolbar__drag-region {
+  z-index: -1;
+  top: 4px;
+  left: 4px;
+  display: block;
+  position: absolute;
+  width: calc(100% - 8px);
+  height: calc(100% - 4px);
   -webkit-app-region: drag;
 }
+
+#app[is-window-maximized]
+  .window-toolbar__drag-region {
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: var(--window-toolbar-height);
+  }
 
 .window-toolbar__content-container {
   display: flex;
   align-items: center;
   width: auto;
   min-width: 100%;
-}
-
-.window-toolbar__content--main {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 12px;
+  flex-shrink: 0;
 }
 
 .window-toolbar__content--main__flex {
+  height: calc(var(--window-toolbar-height) - 8px);
   width: -webkit-fill-available;
+  -webkit-app-region: drag;
 }
+
+#app[is-window-maximized]
+  .window-toolbar__content--main__flex {
+    height: var(--window-toolbar-height);
+  }
 
 .window-toolbar__header-container {
   display: grid !important;
@@ -331,5 +480,26 @@ export default {
 .window-toolbar__item {
   -webkit-app-region: no-drag;
   user-select: none;
+}
+
+.window-controls-group {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin-right: calc(-1 * var(--window-toolbar-padding-right));
+}
+
+.window-controls-group__item {
+  height: var(--window-toolbar-height);
+  width: 46px;
+}
+
+.window-controls-group__item:hover {
+  background-color: var(--highlight-color-4);
+}
+
+.window-controls-group__item:focus {
+  outline: none;
+  background-color: var(--highlight-color-4);
 }
 </style>
