@@ -9,6 +9,7 @@ import router from './router'
 import store from './store'
 import vuetify from './plugins/vuetify'
 import utils from './utils/utils'
+import sharedUtils from './utils/sharedUtils'
 import localizeUtils from './utils/localizeUtils'
 import storeUtils from './utils/storeUtils'
 import { eventHub } from './utils/eventHub'
@@ -30,11 +31,11 @@ import WorkspacesMenu from './components/WorkspacesMenu.vue'
 import TabsMenu from './components/TabsMenu.vue'
 import BasicMenu from './components/BasicMenu.vue'
 import NotificationMenu from './components/NotificationMenu.vue'
+import NotificationCard from './components/NotificationCard.vue'
 import SortableList from './components/SortableList.vue'
 import DirItem from './components/DirItem.vue'
 import DirItemRow from './components/DirItemRow.vue'
 import ContextMenus from './components/ContextMenus.vue'
-import BasicItemCardIterator from './components/BasicItemCardIterator.vue'
 import WorkspaceAreaContent from './components/WorkspaceAreaContent.vue'
 import VirtualWorkspaceAreaContent from './components/VirtualWorkspaceAreaContent.vue'
 import AddressBar from './components/AddressBar.vue'
@@ -42,9 +43,22 @@ import InfoPanel from './components/InfoPanel.vue'
 import GlobalSearch from './components/GlobalSearch.vue'
 import FilterField from './components/FilterField.vue'
 import MediaIterator from './components/MediaIterator.vue'
-import DriveCardsIterator from './components/DriveCardsIterator.vue'
+import ItemCard from './components/ItemCard.vue'
+import ItemCardGrid from './components/ItemCardGrid.vue'
+import DeviceSection from './components/DeviceSection.vue'
+import UserDirsSection from './components/UserDirsSection.vue'
 import UpdatingComponent from './components/UpdatingComponent.vue'
 import SectionSettings from './components/SectionSettings.vue'
+import WindowEffects from './components/WindowEffects.vue'
+import HomeBanner from './components/HomeBanner.vue'
+import FsLocalServerManager from './components/FsLocalServerManager.vue'
+import ShortcutList from './components/ShortcutList.vue'
+import NavigatorTab from './components/NavigatorTab.vue'
+import NavigatorTabBar from './components/NavigatorTabBar.vue'
+import DragSortableList from './components/DragSortableList.vue'
+import SortingHeader from './components/SortingHeader.vue'
+import SortingMenu from './components/SortingMenu.vue'
+import SortingColumnMenu from './components/SortingColumnMenu.vue'
 
 // Register components
 Vue.component('window-toolbar', WindowToolbar)
@@ -62,11 +76,11 @@ Vue.component('workspaces-menu', WorkspacesMenu)
 Vue.component('tabs-menu', TabsMenu)
 Vue.component('basic-menu', BasicMenu)
 Vue.component('notification-menu', NotificationMenu)
+Vue.component('notification-card', NotificationCard)
 Vue.component('sortable-list', SortableList)
 Vue.component('dir-item', DirItem)
 Vue.component('dir-item-row', DirItemRow)
 Vue.component('context-menus', ContextMenus)
-Vue.component('basic-item-card-iterator', BasicItemCardIterator)
 Vue.component('workspace-area-content', WorkspaceAreaContent)
 Vue.component('virtual-workspace-area-content', VirtualWorkspaceAreaContent)
 Vue.component('address-bar', AddressBar)
@@ -74,9 +88,22 @@ Vue.component('info-panel', InfoPanel)
 Vue.component('global-search', GlobalSearch)
 Vue.component('filter-field', FilterField)
 Vue.component('media-iterator', MediaIterator)
-Vue.component('drive-cards-iterator', DriveCardsIterator)
+Vue.component('device-section', DeviceSection)
+Vue.component('user-dirs-section', UserDirsSection)
+Vue.component('item-card', ItemCard)
+Vue.component('item-card-grid', ItemCardGrid)
 Vue.component('updating-component', UpdatingComponent)
 Vue.component('section-settings', SectionSettings)
+Vue.component('window-effects', WindowEffects)
+Vue.component('home-banner', HomeBanner)
+Vue.component('fs-local-server-manager', FsLocalServerManager)
+Vue.component('shortcut-list', ShortcutList)
+Vue.component('navigator-tab', NavigatorTab)
+Vue.component('navigator-tab-bar', NavigatorTabBar)
+Vue.component('drag-sortable-list', DragSortableList)
+Vue.component('sorting-header', SortingHeader)
+Vue.component('sorting-menu', SortingMenu)
+Vue.component('sorting-column-menu', SortingColumnMenu)
 
 const localize = require('./utils/localize')
 require('overlayscrollbars/css/OverlayScrollbars.css')
@@ -84,6 +111,7 @@ require('@mdi/font/css/materialdesignicons.css')
 require('@fortawesome/fontawesome-free/css/all.css')
 
 Vue.prototype.$utils = utils
+Vue.prototype.$sharedUtils = sharedUtils
 Vue.prototype.$localizeUtils = localizeUtils
 Vue.prototype.$storeUtils = storeUtils
 Vue.prototype.$localize = localize
