@@ -4,7 +4,7 @@
 
 const description = {
   name: 'English',
-  locale: 'en'
+  locale: 'en',
 }
 
 const data = {
@@ -26,10 +26,42 @@ const data = {
   tooltip_button_ui_zoom_reset_title: 'Reset zoom level',
   tooltip_button_full_screen_title: 'Full screen',
   tooltip_text_to_copy: 'To copy',
+  language_tags: 'language',
+  ui_scaling_tags: 'ui, scaling, scale, size, zoom, interface, font, magnify',
+  updates_tags: 'updates, version, app',
+  app_properties_tags: 'startup, start, launch, login',
+  window_controls_tags: 'window, controls, minimize, close, restore, hide, tray, memory, ram, app, performance',
+  visual_effects_tags: 'visual, effects, blur, opacity, parallax, overlay, background, transparency, media, color, theme, home, page',
+  theme_tags: 'theme, style, light, dark, color',
+  visual_filters_tags: 'visual, filters, color, contrast, brightness, saturation, image, video, theme',
+  animations_tags: 'animations, home, page',
+  date_time_tags: 'date, time, format',
+  overlays_tags: 'overlays, navigation, panel, drive, storage, letter, ui, elements, visual',
+  ui_elements_tags: 'ui, elements, visual, input, accent, color, timeline, dashboard, page, recent, cards, home, drive, space, indicator, style',
+  home_page_media_banner_tags: 'home, page, media, banner, background, image, video',
+  info_panel_tags: 'info, panel, directory, size, performance, information, details, properties',
+  navigator_tags: 'navigator, dividers, titles, sorting, view, layout, list, grid, item, file, directory, hover, scale, highlight, OCD, ',
+  shortcuts_tags: 'shortcuts',
+  gpu_system_memory_tags: 'gpu, system, memory, ram, performance',
+  workspaces_tags: 'workspaces',
+  tabs_tags: 'tabs, layout',
+  navigator_history_tags: 'navigator, navigation, history',
+  input_navigator_tags: 'input, navigator, click, single, double, mouse, delay, open,',
+  input_elements_tags: 'input, filtering, type, typing, driectory, spellcheck',
+  mouse_buttons_tags: 'mouse, buttons, input, actions',
+  global_search_tags: 'global, search, find, look, filter, scan',
+  data_backup_reminder_tags: 'data, storage, backup',
+  drive_detection_tags: 'drive, usb, storage, disk, device, detection, connection',
+  image_thumbnails_tags: 'image, picture, thumbnails, limit, storage, resources, usage, performance',
+  directory_item_statistics_tags: 'directory, item, statistics, stats, data, storage, usage',
+  app_properties: 'app properties',
+  app_updates: 'app updates',
+  check_updates_now: 'check for updates now',
+  current_version: 'current version',
   text_drive: 'Drive',
   text_drives: 'Drives',
   text_global_search_title: 'Global search',
-  text_system_directories: 'System directories',
+  text_user_directories: 'User directories',
   text_directory: 'directory',
   text_file: 'file',
   text_file_symlink: 'file-symlink',
@@ -56,10 +88,33 @@ const data = {
   text_cancel: 'cancel',
   text_clear_list_message: 'item list has been cleared',
   text_workspace_added: 'workspace has been added',
-  text_allow_global_search_message: 'Global search lets you quickly find any file / directory on your computer.<br>It periodically scans your drives and generates data files (maps) so it can find your files quickly.<br>The files that were created after the last scan will not be found since they won\'t be in the data files, so it\'s recommended to choose a more frequent auto scan period.'
+  text_allow_global_search_message: 'Global search lets you quickly find any file / directory on your computer.<br>It periodically scans your drives and generates data files (maps) so it can find your files quickly.<br>The files that were created after the last scan will not be found since they won\'t be in the data files, so it\'s recommended to choose a more frequent auto scan period.',
+  min: 'min',
+  sec: 'sec',
+  media_info: 'Media info',
+  stream: 'stream',
+  bits_per_raw_sample: 'bits per raw sample',
+  color_space: 'color space',
+  pix_fmt: 'pixel format',
+  chroma_location: 'chroma location',
+  bit_rate: 'bit rate',
+  duration: 'duration',
+  encoder: 'encoder',
+  channel_layout: 'channel layout',
+  channels: 'channels',
+  sample_rate: 'sample rate',
+  avg_frame_rate: 'frame rate',
+  codec_name: 'codec name',
+  codec_tag_string: 'codec tag string',
+  width: 'width',
+  height: 'height',
+  display_aspect_ratio: 'display aspect ratio',
+  has_b_frames: 'has b frames',
+  profile: 'profile',
+  closed_captions: 'closed captions',
 }
 
 export {
   description,
-  data
+  data,
 }
