@@ -5,29 +5,35 @@ Copyright © 2021 - present Aleksey Hoffman. All rights reserved.
 
 <template>
   <div id="navigator-route">
-    <context-menus/>
-    <info-panel/>
-    <!-- content-area -->
+    <info-panel />
     <div
-      :clipboard-toolbar-visible="clipboardToolbarIsVisible"
-      class="content-area"
       id="content-area"
+      :clipboard-toolbar-visible="clipboardToolbarIsVisible"
+      class="content-area fade-in-1s"
     >
-      <!-- workspace-area -->
-      <div class="workspace__area__container" columns_1>
+      <div
+        class="workspace__area__container"
+        columns_1
+      >
         <div class="workspace__area">
           <!-- workspace-area::toolbar -->
-          <v-layout align-center class="workspace-area__header">
+          <v-layout
+            align-center
+            class="workspace-area__header"
+          >
             <!-- button::current-directory-context-menu -->
             <v-tooltip bottom>
-              <template v-slot:activator="{ on }">
+              <template #activator="{ on }">
                 <v-btn
+                  icon
+                  class="workspace-area-toolbar__item"
                   v-on="on"
                   @click="toggleCurrentDirContextMenu($event)"
-                  icon
-                  class="action-toolbar__item"
                 >
-                  <v-icon class="action-toolbar__icon" size="32px">
+                  <v-icon
+                    class="action-toolbar__icon"
+                    size="32px"
+                  >
                     mdi-menu-down
                   </v-icon>
                 </v-btn>
@@ -37,14 +43,17 @@ Copyright © 2021 - present Aleksey Hoffman. All rights reserved.
 
             <!-- button::go-backward-in-history -->
             <v-tooltip bottom>
-              <template v-slot:activator="{ on }">
+              <template #activator="{ on }">
                 <v-btn
+                  icon
+                  class="workspace-area-toolbar__item"
                   v-on="on"
                   @click="$store.dispatch('LOAD_PREVIOUS_HISTORY_PATH')"
-                  icon
-                  class="action-toolbar__item"
                 >
-                  <v-icon class="action-toolbar__icon" size="20px">
+                  <v-icon
+                    class="action-toolbar__icon"
+                    size="20px"
+                  >
                     mdi-arrow-left
                   </v-icon>
                 </v-btn>
@@ -54,14 +63,17 @@ Copyright © 2021 - present Aleksey Hoffman. All rights reserved.
 
             <!-- button::go-forward-in-history -->
             <v-tooltip bottom>
-              <template v-slot:activator="{ on }">
+              <template #activator="{ on }">
                 <v-btn
+                  icon
+                  class="workspace-area-toolbar__item"
                   v-on="on"
                   @click="$store.dispatch('LOAD_NEXT_HISTORY_PATH')"
-                  icon
-                  class="action-toolbar__item"
                 >
-                  <v-icon class="action-toolbar__icon" size="20px">
+                  <v-icon
+                    class="action-toolbar__icon"
+                    size="20px"
+                  >
                     mdi-arrow-right
                   </v-icon>
                 </v-btn>
@@ -71,14 +83,17 @@ Copyright © 2021 - present Aleksey Hoffman. All rights reserved.
 
             <!-- button::go-up-directory -->
             <v-tooltip bottom>
-              <template v-slot:activator="{ on }">
+              <template #activator="{ on }">
                 <v-btn
+                  icon
+                  class="workspace-area-toolbar__item"
                   v-on="on"
                   @click="$store.dispatch('GO_UP_DIRECTORY')"
-                  icon
-                  class="action-toolbar__item"
                 >
-                  <v-icon class="action-toolbar__icon" size="20px">
+                  <v-icon
+                    class="action-toolbar__icon"
+                    size="20px"
+                  >
                     mdi-arrow-up
                   </v-icon>
                 </v-btn>
@@ -88,17 +103,20 @@ Copyright © 2021 - present Aleksey Hoffman. All rights reserved.
 
             <!-- button::reload -->
             <v-tooltip bottom>
-              <template v-slot:activator="{ on }">
+              <template #activator="{ on }">
                 <v-btn
-                  v-on="on"
-                  @click="$store.dispatch('LOAD_DIR', {
-                    path: currentDir.path,
-                    scrollTop: false
-                  })"
                   icon
-                  class="action-toolbar__item"
+                  class="workspace-area-toolbar__item"
+                  v-on="on"
+                  @click="$store.dispatch('RELOAD_DIR', {
+                    scrollTop: false,
+                    emitNotification: true
+                  })"
                 >
-                  <v-icon class="action-toolbar__icon" size="20px">
+                  <v-icon
+                    class="action-toolbar__icon"
+                    size="20px"
+                  >
                     mdi-refresh
                   </v-icon>
                 </v-btn>
@@ -106,27 +124,34 @@ Copyright © 2021 - present Aleksey Hoffman. All rights reserved.
               <span>Reload current directory</span>
             </v-tooltip>
 
-            <address-bar/>
-            <v-spacer/>
-            <filter-field/>
+            <address-bar />
+            <v-spacer />
+            <filter-field />
           </v-layout>
 
-          <global-search v-show="globalSearchWidget"/>
-          <div v-if="filterQuery.length > 0" class="content__description">
+          <sorting-header v-if="navigatorSortingElementDisplayType === 'bar'" />
+
+          <div
+            v-if="filterQuery.length > 0"
+            class="content__description"
+          >
             <v-btn
-              @click="filterQuery = ''"
               small
               class="mt-5 mx-6 button-1"
+              @click="filterQuery = ''"
             >
               <v-icon
                 size="16px"
                 class="mr-2"
-              >mdi-backspace-outline
+              >
+                mdi-backspace-outline
               </v-icon>
               Clear filter
             </v-btn>
           </div>
-          <workspace-area-content/>
+
+          <global-search v-show="globalSearchWidget" />
+          <workspace-area-content />
         </div>
       </div>
     </div>
@@ -134,8 +159,8 @@ Copyright © 2021 - present Aleksey Hoffman. All rights reserved.
 </template>
 
 <script>
-import { mapFields } from 'vuex-map-fields'
-import { mapGetters } from 'vuex'
+import {mapFields} from 'vuex-map-fields'
+import {mapGetters} from 'vuex'
 
 export default {
   name: 'navigator',
@@ -144,29 +169,15 @@ export default {
       processedDirItemsIndexes: [],
       dirItemAwaitsSecondClick: false,
       dirItemAwaitsSecondClickTimeout: null,
-      dirItemSecondClickDelay: 500
+      dirItemSecondClickDelay: 500,
     }
   },
   beforeRouteLeave (to, from, next) {
-    this.$eventHub.$emit('app:method', {
-      method: 'setRouteScrollPosition',
-      params: {
-        toRoute: to,
-        fromRoute: from
-      }
+    this.$store.dispatch('SAVE_ROUTE_SCROLL_POSITION', {
+      toRoute: to,
+      fromRoute: from,
     })
     next()
-  },
-  async mounted () {
-    this.$nextTick(() => {
-      this.$store.dispatch('ROUTE_MOUNTED_HOOK_CALLBACK', {
-        route: 'navigator'
-      })
-      this.navigatorRouteIsLoaded = true
-    })
-  },
-  beforeDestroy () {
-    this.navigatorRouteIsLoaded = false
   },
   watch: {
     // dirItemsInfoIsFetched (value) {
@@ -182,20 +193,37 @@ export default {
       // deselected (and the current dir gets selected automatically)
       // so that the context menu does not remain displayed for the
       // deselected items
-      if (this.onlyCurrentDirIsSelected) {
+      if (this.isOnlyCurrentDirItemSelected) {
         this.contextMenus.dirItem.value = false
       }
     },
     'navigatorView.selectedDirItems' (value) {
       this.contextMenus.dirItem.value = false
-    }
+    },
+  },
+  activated () {
+    this.$store.dispatch('ROUTE_ACTIVATED_HOOK_CALLBACK', {
+      route: 'navigator',
+    })
+  },
+  async mounted () {
+    this.setSortingTypes()
+    this.$nextTick(() => {
+      this.$store.dispatch('ROUTE_MOUNTED_HOOK_CALLBACK', {
+        route: 'navigator',
+      })
+      this.navigatorRouteIsLoaded = true
+    })
+  },
+  beforeDestroy () {
+    this.navigatorRouteIsLoaded = false
   },
   computed: {
     ...mapGetters([
       'selectedWorkspace',
       'selectedDirItems',
       'clipboardToolbarIsVisible',
-      'onlyCurrentDirIsSelected'
+      'isOnlyCurrentDirItemSelected',
     ]),
     ...mapFields({
       focusedField: 'focusedField',
@@ -210,8 +238,11 @@ export default {
       currentDir: 'navigatorView.currentDir',
       dirItemsInfoIsFetched: 'navigatorView.dirItemsInfoIsFetched',
       pinnedItems: 'storageData.pinned.items',
-      protectedItems: 'storageData.protected.items'
-    })
+      sortingOrder: 'sorting.order',
+      selectedSortingType: 'sorting.selectedType',
+      sortingTypes: 'sorting.types',
+      navigatorSortingElementDisplayType: 'storageData.settings.navigator.sorting.elementDisplayType',
+    }),
   },
   methods: {
     toggleCurrentDirContextMenu (payload) {
@@ -221,11 +252,28 @@ export default {
           this.$store.dispatch('SET_CONTEXT_MENU', {
             value: 'toggle',
             x: payload.clientX,
-            y: payload.clientY
+            y: payload.clientY,
           })
         })
-    }
-  }
+    },
+    setSortingTypes () {
+      let sortingTypesClone = this.sortingTypes.map(displayedType => {
+        displayedType.onClick = (item) => this.handleSortHeaderItemClick(item)
+        displayedType.type = 'sort'
+        return displayedType
+      })
+
+      this.sortingTypes = sortingTypesClone
+    },
+    handleSortHeaderItemClick (item) {
+      if (this.selectedSortingType.name === item.name) {
+        this.$store.dispatch('TOGGLE_SORTING_ORDER')
+      }
+      else {
+        this.$store.dispatch('SET_SORTING_TYPE', item)
+      }
+    },
+  },
 }
 </script>
 
@@ -233,33 +281,28 @@ export default {
 .workspace__area__container[columns_2] {
   display: grid;
   grid-template-columns: 1fr 1fr;
-  /* gap: 20px; */
 }
 
 .workspace__area__container[columns_1] {
   display: grid;
   grid-template-columns: 1fr;
-  /* gap: 20px; */
 }
 
 .workspace__area:nth-child(2) {
   border-left: 1px solid var(--dir-item-card-border);
 }
 
-.workspace__area__name {
-}
-
 .workspace__area {
   display: grid;
-  grid-template-rows: 48px 1fr;
-  /* gap: 20px; */
+  grid-template-rows: 48px 38px 1fr;
 
 }
 
 .workspace-area__header {
-  /* padding: 8px 24px;
-  background-color: rgb(255, 255, 255, 0.02); */
-
+  display: flex;
+  align-items: center;
+  gap: var(--toolbar-item-gap);
+  height: var(--workspace-area-toolbar-height);
   margin: 0px;
   margin-bottom: 0px;
   padding: 0 8px;
@@ -278,10 +321,6 @@ export default {
     var(--action-toolbar-height) -
     var(--workspace-area-toolbar-height)
   );
-}
-
-.info-panel__preview-container {
-
 }
 
 .info-panel__media-preview {
@@ -317,18 +356,4 @@ export default {
     .workspace-area__content {
       margin: 0px;
     }
-
-#navigator-route
-  .content-area[clipboard-toolbar-visible] {
-    /* height: calc(100vh - var(--window-toolbar-height) - var(--action-toolbar-height) - 32px); */
-    /* padding-bottom: 32px; */
-    overflow: hidden;
-  }
-
-#navigator-route
-  .content-area[clipboard-toolbar-visible]
-    .workspace-area__content {
-      padding-bottom: 132px;
-    }
-
 </style>
